% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{arrange_by}
\alias{arrange_by}
\title{Sort vector by another vector}
\usage{
arrange_by(x, by)
}
\arguments{
\item{x}{Vector to be ordered.}

\item{by}{Reference vector which \code{x} will be ordered by.}
}
\value{
A permutation of \code{x}.
}
\description{
Arranges a vector \code{x} by the order of another vector \code{by}.
}
\details{
Note that this function significantly differs from \code{\link[base:sort_by]{base::sort_by()}}.
}
\examples{
library(magrittr)

# generate 100 random letters
random_letters <-
  letters |>
  magrittr::extract(sample.int(n = 26L,
                               size = 100L,
                               replace = TRUE)) \%T>\%
  print()

# sort the random letters alphabetically
random_letters |> pal::arrange_by(by = letters)
}
