% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{as_comment_string}
\alias{as_comment_string}
\title{Assemble an (R) comment string of the desired line width}
\usage{
as_comment_string(..., line_width = 160L, comment_prefix = "# ")
}
\arguments{
\item{...}{Comment lines. A character vector or something coercible to.}

\item{line_width}{Maximum character width at which to wrap lines. An integer scalar.}

\item{comment_prefix}{Character sequence that indicates the start of a comment. A character scalar.}
}
\value{
A character scalar.
}
\description{
Takes a vector of text lines, wraps them at the specified line width, prefixes them with comment markers and returns the result as a single string.
}
\examples{
pal::as_comment_string(glue::glue("Copyright (C) {format(Sys.Date(), '\%Y')} Santa Clause"),
                       "No presents without code.") |>
cat()
}
