% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{as_env_var_name}
\alias{as_env_var_name}
\title{Convert to a valid environment variable name}
\usage{
as_env_var_name(...)
}
\arguments{
\item{...}{Input to convert to an environment variable name. A character vector or something coercible to.}
}
\value{
A character scalar.
}
\description{
Convert to a valid environment variable name
}
\examples{
pal::as_env_var_name("foo.bar")
pal::as_env_var_name("/oh/my/goodness")
pal::as_env_var_name("1.2", "some title")
}
\seealso{
Other string functions:
\code{\link{as_chr}()},
\code{\link{as_comment_str}()},
\code{\link{as_line_feed_chr}()},
\code{\link{as_str}()},
\code{\link{capitalize_first}()},
\code{\link{dsv_colnames}()},
\code{\link{enum_str}()},
\code{\link{escape_lf}()},
\code{\link{fuse_regex}()},
\code{\link{phrase_nr}()},
\code{\link{sentenceify}()},
\code{\link{wrap_chr}()}
}
\concept{string}
