% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{as_line_feed_chr}
\alias{as_line_feed_chr}
\title{Convert control character sequence name to actual character sequence}
\usage{
as_line_feed_chr(eol = c("LF", "CRLF", "CR", "LFCR"))
}
\arguments{
\item{eol}{End of line (EOL) control character sequence. One of
\itemize{
\item \code{"LF"} for the line feed (LF) character (\code{"\\n"}). The standard on Unix and Unix-like systems (Linux, macOS, *BSD, etc.) and the \strong{default}.
\item \code{"CRLF"} for the carriage return + line feed (CR+LF) character sequence (\code{"\\r\\n"}). The standard on Microsoft Windows, DOS and some other systems.
\item \code{"CR"} for the carriage return (CR) character (\code{"\\r"}). The standard on classic Mac OS and some other antiquated systems.
\item \code{"LFCR"} for the line feed + carriage return (LF+CR) character sequence (\code{"\\n\\r"}). The standard on RISC OS and some other exotic systems.
}}
}
\value{
A character scalar.
}
\description{
Convert control character sequence name to actual character sequence
}
\seealso{
Other string functions:
\code{\link{as_chr}()},
\code{\link{as_comment_string}()},
\code{\link{as_string}()},
\code{\link{dsv_colnames}()},
\code{\link{escape_lf}()},
\code{\link{fuse_regex}()},
\code{\link{prettify_nr}()},
\code{\link{prose_ls}()},
\code{\link{sentenceify}()},
\code{\link{wrap_chr}()}
}
\concept{string}
