% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{as_md_list}
\alias{as_md_list}
\title{Convert a character vector to a Markdown list}
\usage{
as_md_list(
  x,
  type = c("unordered", "ordered", "ordered_roman"),
  tight = TRUE,
  indent_lvl = 0L,
  wrap = NULL
)
}
\arguments{
\item{x}{The character vector to be converted to a Markdown list.}

\item{type}{The Markdown list type. One of
\itemize{
\item \code{"unordered"} for an unordered aka \href{https://pandoc.org/MANUAL.html#bullet-lists}{bullet list}. Corresponds to \verb{<ul>} in HTML.
\item \code{"ordered"} for an ordered aka \href{https://pandoc.org/MANUAL.html#ordered-lists}{numbered list}. Corresponds to \verb{<ol>} in HTML.
\item \code{"ordered_roman"} for a variation of an ordered/numbered list with uppercase roman numerals instead of Arabic numerals as list markers.
}}

\item{tight}{Whether or not to add additional spacing between list items.}

\item{indent_lvl}{The level of indentation of the resulting Markdown list. For each level, four additional spaces are added in front of every list item. An
integer scalar.}

\item{wrap}{An optional string to wrap the list items in.}
}
\value{
A character scalar.
}
\description{
Convenience wrapper around \code{\link[pander:pandoc.list.return]{pander::pandoc.list.return()}} to convert a character vector (or something coercible to) to a \href{https://pandoc.org/MANUAL.html#lists}{Markdown list}.
}
\examples{
rownames(mtcars) |> pal::as_md_list() |> cat()
}
\seealso{
Other (Pandoc) Markdown functions:
\code{\link{pipe_table}()}
}
\concept{md}
