% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{as_md_val_list}
\alias{as_md_val_list}
\title{List values as a Markdown list}
\usage{
as_md_val_list(...)
}
\arguments{
\item{...}{\R objects to list.}
}
\value{
A character scalar.
}
\description{
Generates a Markdown list of the given values formatted as \href{https://pandoc.org/MANUAL.html#verbatim}{verbatim}. Character values are additionall wrapped in
double quotes.
}
\examples{
list(1L, 2.2, "other") |>
  pal::as_md_val_list() |>
  cat()
  
# note that values are flattened before listing them, so this yields the same list
list(list(list(1L, list(2.2)))) |>
  pal::as_md_val_list() |>
  cat()
}
\seealso{
Other (Pandoc) Markdown functions:
\code{\link{as_md_list}()},
\code{\link{pipe_table}()},
\code{\link{strip_md_footnotes}()},
\code{\link{strip_md}()}
}
\concept{md}
