% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{as_md_vals}
\alias{as_md_vals}
\title{Format values as verbatim Markdown}
\usage{
as_md_vals(...)
}
\arguments{
\item{...}{Values to be formatted. One or more \R objects.}
}
\value{
A character vector.
}
\description{
Converts the given values to a character vector, formatted as \href{https://pandoc.org/MANUAL.html#verbatim}{verbatim Markdown}. Character values are additionally
wrapped in double quotes. Metadata like \link[base:typeof]{types} or \link[base:attributes]{attributes} of the input is \emph{not} represented in the output, only its
\emph{values}.
}
\examples{
list(1L, 2.2, "other") |>
  pal::as_md_vals()
  
# note that values are flattened before further processing them, so this yields the same result
list(list(list(1L, list(2.2), list(list("other"))))) |>
  pal::as_md_vals()
}
\seealso{
Other (Pandoc) Markdown functions:
\code{\link{as_md_list}()},
\code{\link{as_md_val_list}()},
\code{\link{md_verb}()},
\code{\link{pipe_table}()},
\code{\link{strip_md_footnotes}()},
\code{\link{strip_md}()}
}
\concept{md}
