% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{as_str}
\alias{as_str}
\title{Convert to a character scalar (aka string)}
\usage{
as_str(..., sep = "", rm_na = FALSE)
}
\arguments{
\item{...}{\R objects to be assembled to a single string. \link[rlang:dyn-dots]{Dynamic dots} are supported.}

\item{sep}{Separator to delimit \code{...}. Defaults to none (\code{""}).}

\item{rm_na}{Exclude missing values. If \code{FALSE}, missing values will be represented as \code{"NA"} in the resulting string.}
}
\value{
A character scalar.
}
\description{
Similar to \code{\link[base:paste]{paste0(..., collapse = "")}}, but \emph{recursively} converts its inputs to type character.
}
\examples{
library(magrittr)

input <-
  paste0(2:4,
         collapse = ", ") |>
  purrr::map(rep,
             times = 20) \%>\%
  list(c("This is a glut of ", "meaningless numbers: "), .)

# while this just converts `input` in a lazy way...
paste0(input,
       collapse = "")

# ...this one works harder
pal::as_str(input)
}
\seealso{
Other string functions:
\code{\link{as_chr}()},
\code{\link{as_comment_str}()},
\code{\link{as_line_feed_chr}()},
\code{\link{capitalize_first}()},
\code{\link{dsv_colnames}()},
\code{\link{enum_str}()},
\code{\link{escape_lf}()},
\code{\link{fuse_regex}()},
\code{\link{prettify_nr}()},
\code{\link{prose_ls}()},
\code{\link{sentenceify}()},
\code{\link{wrap_chr}()}
}
\concept{string}
