% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{assert_cli}
\alias{assert_cli}
\title{Assert CLI tool is available}
\usage{
assert_cli(
  cmd,
  error_msg =
    paste0("The {.strong {cmd}} executable is required but couldn't be found on ",
    "system's {.href [PATH](https://en.wikipedia.org/wiki/PATH_(variable))}."),
  force_which = FALSE
)
}
\arguments{
\item{cmd}{System command to invoke the CLI tool. A character scalar.}

\item{error_msg}{Error message to print in case \code{cmd} is not found on system's \code{PATH}. Supports cli's \link[cli:inline-markup]{inline markup}. A character scalar.}

\item{force_which}{If set to \code{TRUE}, \code{\link[base:Sys.which]{base::Sys.which()}}, which relies on the system command \code{which}, will be used instead of \code{command -v} to determine the
availability of \code{cmd} on Unix-like systems. On Windows, \code{base::Sys.which()} is used in any case. \code{command -v} is
\href{https://unix.stackexchange.com/q/85249/201803}{generally recommended for bourne-like shells} and therefore is the default on Linux, macOS and other
\href{https://en.wikipedia.org/wiki/Unix-like}{Unixes}.}
}
\value{
If the CLI tool is available on the system's \code{PATH}, its filesystem path, invisibly. Otherwise, an error is thrown.
}
\description{
Asserts that a \href{https://en.wikipedia.org/wiki/Command-line_interface}{command-line interface (CLI)} tool is found on the system's
\href{https://en.wikipedia.org/wiki/PATH_(variable)}{\code{PATH}} and returns the executable's filesystem path.
}
\examples{
pal::assert_cli("Rscript")

cmd <- ifelse(xfun::is_windows(), "pandoc.exe", "pandoc")
pal::assert_cli(cmd)
}
\seealso{
Other system interaction functions:
\code{\link{path_script}()},
\code{\link{test_cli}()}

Other functions extending the checkmate package:
\code{\link{assert_class_any}()},
\code{\link{assert_df_or_tibble}()},
\code{\link{assert_inf_count}()},
\code{\link{assert_mime_type}()},
\code{\link{assert_pkg}()}
}
\concept{checkmate}
\concept{sys}
