% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{assert_cols}
\alias{assert_cols}
\title{Assert data frame columns}
\usage{
assert_cols(
  data,
  cols = NULL,
  strict = FALSE,
  obj_name = checkmate::vname(data)
)
}
\arguments{
\item{data}{Data frame to check.}

\item{cols}{Column names which must be present in \code{data}. A character vector or \code{NULL}.}

\item{strict}{Whether additional columns not specified in \code{cols} are allowed in \code{data}.}

\item{obj_name}{Name of the checked object to print in error messages.}
}
\value{
\code{data}, invisibly.
}
\description{
Asserts that a data frame contains the specified columns.
}
\examples{
pal::assert_cols(data = mtcars,
                 cols = c("mpg", "disp"))

try(
  pal::assert_cols(data = mtcars,
                   cols = c("mpg", "display"))
)

try(
  pal::assert_cols(data = mtcars,
                   cols = c("mpg", "disp"),
                   strict = TRUE)
)

try(
  pal::assert_cols(data = mtcars,
                   strict = TRUE)
)
}
\seealso{
Other data frame / tibble functions:
\code{\link{is_equal_df}()},
\code{\link{reduce_df_list}()}
}
\concept{df}
