% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{assert_df_or_tibble}
\alias{assert_df_or_tibble}
\title{Assert object is data frame or tibble (extension)}
\usage{
assert_df_or_tibble(x, name = "x")
}
\arguments{
\item{x}{\R object to test.}

\item{name}{Name of the checked object to print in error message in case the assertion fails. A character scalar.}
}
\value{
\code{x}, invisibly.
}
\description{
Asserts that an object is \code{pkgsnip::param_lbl(id = "df_or_tibble", as_sentence = FALSE)}.
}
\examples{
tibble::tibble() |> pal::assert_df_or_tibble()

try(
  matrix() |> pal::assert_df_or_tibble()
)
}
\seealso{
Other functions extending the checkmate package:
\code{\link{assert_class_any}()},
\code{\link{assert_cli}()},
\code{\link{assert_inf_count}()},
\code{\link{assert_mime_type}()},
\code{\link{assert_pkg}()}
}
\concept{checkmate}
