% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{assert_inf_count}
\alias{assert_inf_count}
\title{Assert count or \code{Inf}}
\usage{
assert_inf_count(x, ...)
}
\arguments{
\item{x}{Object to check.}

\item{...}{Further arguments passed on to \code{\link[checkmate:checkCount]{checkmate::assert_count()}}.}
}
\value{
\code{x}, invisibly.
}
\description{
Asserts that \code{x} is either a \link[checkmate:checkCount]{count} or \link[=is.infinite]{positive infinity} (\code{Inf}).
}
\examples{
pal::assert_inf_count(1001L)
pal::assert_inf_count(1)
pal::assert_inf_count(0.0)
pal::assert_inf_count(Inf)

try(
  pal::assert_inf_count(-1)
)
try(
  pal::assert_inf_count(1.5)
)
try(
  pal::assert_inf_count(-Inf)
)
}
\seealso{
Other functions extending the checkmate package:
\code{\link{assert_class_any}()},
\code{\link{assert_cli}()},
\code{\link{assert_df_tbl}()},
\code{\link{assert_mime_type}()},
\code{\link{assert_pkg}()}
}
\concept{checkmate}
