% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{capitalize_1st}
\alias{capitalize_1st}
\title{Capitalize first letter}
\usage{
capitalize_1st(x)
}
\arguments{
\item{x}{Input of which to capitalize the first letter, typically a character vector.}
}
\value{
A character vector of the same length as \code{x}.
}
\description{
Converts the input to a character vector, with the first letter of each element uppercased.
}
\examples{
pal::capitalize_1st(c("one", "Two", "THREE"))
}
\seealso{
Other string functions:
\code{\link{as_chr}()},
\code{\link{as_comment_string}()},
\code{\link{as_line_feed_chr}()},
\code{\link{as_string}()},
\code{\link{dsv_colnames}()},
\code{\link{escape_lf}()},
\code{\link{fuse_regex}()},
\code{\link{prettify_nr}()},
\code{\link{prose_ls}()},
\code{\link{sentenceify}()},
\code{\link{wrap_chr}()}
}
\concept{string}
