% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{capitalize_first}
\alias{capitalize_first}
\title{Capitalize first letter}
\usage{
capitalize_first(x)
}
\arguments{
\item{x}{Input of which to capitalize the first letter, typically a character vector.}
}
\value{
A character vector of the same length as \code{x}.
}
\description{
Converts the input to a character vector, with the first letter of each element uppercased.
}
\examples{
pal::capitalize_first(c("one", "Two", "tHREE"))
}
\seealso{
Other string functions:
\code{\link{as_chr}()},
\code{\link{as_comment_str}()},
\code{\link{as_line_feed_chr}()},
\code{\link{as_str}()},
\code{\link{dsv_colnames}()},
\code{\link{enum_str}()},
\code{\link{escape_lf}()},
\code{\link{fuse_regex}()},
\code{\link{prettify_nr}()},
\code{\link{prose_ls}()},
\code{\link{sentenceify}()},
\code{\link{wrap_chr}()}
}
\concept{string}
