% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-package.R
\docType{package}
\name{chores-package}
\alias{chores}
\alias{chores-package}
\title{chores: A Collection of Large Language Model Assistants}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a collection of ergonomic large language model assistants designed to help you complete repetitive, hard-to-automate tasks quickly. After selecting some code, press the keyboard shortcut you've chosen to trigger the package app, select an assistant, and watch your code be rewritten. While the package ships with a number of helpers for R package development, users can create custom helpers just by writing some instructions in a markdown file.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/chores}
  \item \url{https://simonpcouch.github.io/chores/}
  \item Report bugs at \url{https://github.com/simonpcouch/chores/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
