% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{cli_progress_step_quick}
\alias{cli_progress_step_quick}
\title{Quick \href{https://cli.r-lib.org/}{cli} simplified progress message}
\usage{
cli_progress_step_quick(
  msg,
  spinner = FALSE,
  class = if (!spinner) ".alert-info",
  current = TRUE,
  .auto_close = TRUE,
  .envir = parent.frame()
)
}
\arguments{
\item{msg}{Message to show. It may contain glue substitution and cli
styling. It can be updated via \code{\link[cli:cli_progress_update]{cli_progress_update()}}, as usual.
It is style as a cli info alert (see \code{\link[cli:cli_alert_info]{cli_alert_info()}}).}

\item{spinner}{Whether to show a spinner at the beginning of the line.
To make the spinner spin, you'll need to call \code{cli_progress_update()}
regularly.}

\item{class}{cli class to add to the message. By default there is no
class for steps with a spinner.}

\item{current}{Whether to use this progress bar as the current progress
bar of the calling function. See more at 'The current progress bar'
below.}

\item{.auto_close}{Whether to terminate the progress bar when the
calling function (or the one with execution environment in \code{.envir}
exits. (Auto termination does not work for progress bars created
from the global environment, e.g. from a script.)}

\item{.envir}{The environment to use for auto-termination and for glue
substitution. It is also used to find and set the current progress bar.}
}
\value{
The cli progress bar \code{id} as a character scalar, invisibly.
}
\description{
Version of \code{\link[cli:cli_progress_step]{cli::cli_progress_step()}} that for the messages on (un)successful termination simply appends \strong{done}/\strong{failed} to \code{msg}.
}
\examples{
pal::cli_progress_step_quick(msg = "Doing my thing")
Sys.sleep(1)
cli::cli_progress_done()
}
\seealso{
Other functions extending the cli package:
\code{\link{cli_process_expr}()},
\code{\link{cli_qty_lgl}()}
}
\concept{cli}
