% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{daff_diff}
\alias{daff_diff}
\title{Determine the differences between two data frames/tibbles in tabular diff format}
\usage{
daff_diff(
  x,
  y,
  ignore_order = FALSE,
  convert = FALSE,
  ids = NULL,
  ask = TRUE,
  bypass_rstudio_viewer = FALSE,
  change_msg = NULL,
  ask_msg = "Do you wish to display the changes?",
  caption = NULL,
  ...
)
}
\arguments{
\item{x}{The data frame / tibble to check for changes.}

\item{y}{The data frame / tibble that \code{x} should be checked against, i.e. the reference.}

\item{ignore_order}{Whether or not to ignore the order of columns and rows.}

\item{convert}{Whether or not to convert similar column types. Currently, this will convert factor to character and integer to double.}

\item{ids}{A character vector of column names that make up a primary key, if known. If \code{NULL}, heuristics are used to find a decent key (or a set of decent
keys).}

\item{ask}{Whether to ask interactively if the resulting difference object should be opened (if \code{x} and \code{y} differ). If \code{FALSE}, it will be opened right
away. Only relevant if run \link[base:interactive]{interactively}.}

\item{bypass_rstudio_viewer}{If \code{TRUE}, \code{x} and \code{y} actually differ, and \code{ask} is set to \code{TRUE}, the resulting difference object will be
opened in the system's default web browser instead of RStudio's built-in viewer. Only relevant if run within RStudio.}

\item{change_msg}{The message that is displayed on the console in case \code{x} and \code{y} differ. If \code{NULL}, a default message is displayed.}

\item{ask_msg}{The message that is displayed when \code{ask = TRUE}. Ignored if \code{ask = FALSE}.}

\item{caption}{The caption of the rendered difference object. A character scalar. If \code{NULL}, a default caption is generated based on the object names of \code{x}
and \code{y}.}

\item{...}{Further arguments passed on to \code{\link[daff:diff_data]{daff::diff_data()}}, excluding \code{data}, \code{data_ref}, \code{ids}, \code{ordered}, and \code{columns_to_ignore}.}
}
\value{
A difference object which can be rendered later using \code{\link[daff:render_diff]{daff::render_diff()}}, invisibly.
}
\description{
Compares two \link[base:data.frame]{data frames}/\link[tibble:tibble]{tibbles} (or two objects coercible to tibbles like
\link[base:matrix]{matrices}) and offers to inspect any differences in \href{https://paulfitz.github.io/daff-doc/spec.html}{tabular diff format} as rendered HTML.
}
\details{
This function is basically a convenience wrapper combining \code{\link[=is_equal_df]{is_equal_df()}}, \code{\link[daff:diff_data]{daff::diff_data()}} and \code{\link[daff:render_diff]{daff::render_diff()}}. Note that it only compares the
\emph{content} of \code{x} and \code{y}, not their attributes.
}
\examples{
\dontrun{
mtcars \%>\%
  dplyr::mutate(dplyr::across(c(cyl, gear),
                              ~ dplyr::if_else(. > 4, . * 2, .))) \%>\%
  daff_diff(mtcars)}
}
\seealso{
Data frame / Tibble functions
\code{\link{all_cols}()},
\code{\link{is_equal_df}()},
\code{\link{open_as_tmp_spreadsheet}()}
}
\concept{tibble}
