% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{deploy_pkgdown_site}
\alias{deploy_pkgdown_site}
\title{Deploy pkgdown site to local Git folder}
\usage{
deploy_pkgdown_site(
  pkg_path = ".",
  to_path,
  use_dev_build = NULL,
  clean_to_path = TRUE,
  commit_msg = paste0("auto-deploy pkgdown site for ", desc_value(file = pkg_path, key
    = "Package"))
)
}
\arguments{
\item{pkg_path}{Path to the \R package of which the pkgdown website files are to be deployed.}

\item{to_path}{Path to the Git (sub)folder to which the pkgdown website files are to be deployed.}

\item{use_dev_build}{Whether to deploy the development build of the pkgdown website files.}

\item{clean_to_path}{Whether to completely wipe \code{to_path} before deploying the new pkgdown website files. Setting this \code{TRUE} ensures there are no obsolete
files left over from previous deployments.}

\item{commit_msg}{The Git commit message used for the deployment.}
}
\value{
A vector of paths to the deployed files/folders, invisibly.
}
\description{
Copies the static \link[pkgdown:pkgdown-package]{pkgdown} website files to another local Git folder, then stages, commits and pushes the changes. Use
\code{\link[pkgdown:build_site]{pkgdown::build_site()}} before running this function in order to create the website files.

Use this function with \strong{caution} since by default it \strong{completely wipes the \code{to_path} directory}!
}
