% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{desc_dep_vrsn}
\alias{desc_dep_vrsn}
\title{Get dependency version from \code{DESCRIPTION} file}
\usage{
desc_dep_vrsn(
  pkg,
  file = ".",
  types = c("Imports", "Depends", "Suggests", "Enhances", "LinkingTo")
)
}
\arguments{
\item{pkg}{Package name. A character scalar. Defaults to the name of the calling package.}

\item{file}{DESCRIPTION file to use. By default the DESCRIPTION
file of the current package (i.e. the package the working directory
is part of) is used.}

\item{types}{Dependency types to be considered. If \code{pkg} is listed in multiple dependency types, the maximally required version is returned.}
}
\value{
A \link[base:numeric_version]{numeric version}.
}
\description{
Returns the version of the specified \code{pkg} dependency from a \code{DESCRIPTION} file.
}
\examples{
fs::path_package(package = "dplyr") |> pal::desc_dep_vrsn(pkg = "tibble")
}
\seealso{
Other package \code{DESCRIPTION} functions:
\code{\link{desc_list}()},
\code{\link{desc_url_git}()},
\code{\link{desc_value}()}
}
\concept{desc}
