% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{desc_list}
\alias{desc_list}
\title{Get all \code{DESCRIPTION} file fields as cleaned up list}
\usage{
desc_list(file = ".")
}
\arguments{
\item{file}{DESCRIPTION file to use. By default the DESCRIPTION
file of the current package (i.e. the package the working directory
is part of) is used.}
}
\value{
A list.
}
\description{
Returns all fields from a specific \code{DESCRIPTION} file as a named list with values cleaned up:
\itemize{
\item Whitespaces at the start and end of field values as well as repeated whitespaces within them are removed.
\item Multi-value fields are returned as vectors.
\item The fields \code{Depends}, \code{Imports} and \code{Suggests} are returned as a single data frame named \code{dependencies}.
}
}
\examples{
\dontrun{
pal::desc_list(file = fs::path_package(package = "pal"))
}
}
\seealso{
Other R package functions:
\code{\link{assert_pkg}()},
\code{\link{desc_value}()},
\code{\link{is_pkg_dir}()},
\code{\link{is_pkg_installed}()},
\code{\link{is_pkgdown_dir}()},
\code{\link{ls_pkg}()},
\code{\link{prose_ls_fn_param}()}
}
\concept{rpkgs}
