% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{desc_url_git}
\alias{desc_url_git}
\title{Get the Git repository URL from a \code{DESCRIPTION} file}
\usage{
desc_url_git(file = ".")
}
\arguments{
\item{file}{DESCRIPTION file to use. By default the DESCRIPTION
file of the current package (i.e. the package the working directory
is part of) is used.}
}
\value{
A character scalar.
}
\description{
Returns the first Git repository URL found in the \code{URL} (preferred) or \code{BugReports} fields of a \code{DESCRIPTION} file.
}
\details{
Currently, this function recognizes \href{https://gitlab.com/}{GitLab}, \href{https://github.com/}{GitHub}, \href{https://gitea.com/}{Gitea},
\href{https://codeberg.org/}{Codeberg}, \href{https://pagure.io/}{Pagure}, \href{https://bitbucket.org/}{Bitbucket} and \href{https://sr.ht/}{SourceHut} repository URLs.
}
