% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-addin.R
\name{.init_addin}
\alias{.init_addin}
\title{Run the pal addin}
\usage{
.init_addin()
}
\value{
\code{NULL}, invisibly. Called for the side effect of launching the pal addin
and interfacing with selected text.
}
\description{
The pal addin allows users to interactively select a pal to
interface with the current selection. \strong{This function is not
intended to be interfaced with in regular usage of the package.}
To launch the pal addin in RStudio, navigate to \code{Addins > Pal}
and/or register the addin with a shortcut via
\verb{Tools > Modify Keyboard Shortcuts > Search "Pal"}--we suggest \code{Ctrl+Alt+P}
(or \code{Ctrl+Cmd+P} on macOS).
}
