% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{.pal_color-class}
\alias{.pal_color-class}
\title{An S4 class to save colors.}
\description{
An S4 class to save colors.
}
\section{Slots}{

\describe{
\item{\code{color}}{character.}

\item{\code{distance}}{numeric.}
}}

