% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{.pal_global-class}
\alias{.pal_global-class}
\title{An S4 class to represent the global parameters of pal.}
\description{
An S4 class to represent the global parameters of pal.
}
\section{Slots}{

\describe{
\item{\code{catalogs}}{character vector; Collections of palettes that
are used in the search. The palette collections available in pal package are: \cr
\itemize{
\item cpt-city: http://soliton.vm.bytemark.co.uk/pub/cpt-city/index.html \cr
\item color-hex: https://www.color-hex.com/color-palettes/ \cr
\item paletteer: https://github.com/EmilHvitfeldt/paletteer \cr
}}

\item{\code{type}}{character vector; Type of palette. The type: \cr
\itemize{
\item 'continuous' \cr
\item 'discrete' \cr
\item 'diverging' \cr
\item 'qualitative' \cr
\item 'sequential' \cr
}

The types: diverging, qualitative and sequential
only are supported by the 'paletteer'catalog.}
}}

