% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{.pal_n-class}
\alias{.pal_n-class}
\title{An S4 class to represent the color complexity.}
\description{
An S4 class to represent the color complexity.
}
\section{Slots}{

\describe{
\item{\code{min}}{numeric; min number of colors.}

\item{\code{max}}{numeric; max number of colors.}
}}

