% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.stash_last_pal}
\alias{.stash_last_pal}
\title{Save most recent results to search path}
\usage{
.stash_last_pal(x)
}
\arguments{
\item{x}{A pal.}
}
\value{
NULL, invisibly.
}
\description{
Save most recent results to search path
}
\details{
The function will assign \code{x} to \code{.last_pal} and put it in
the search path.
}
\keyword{internal}
