% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{dsv_colnames}
\alias{dsv_colnames}
\title{Get column names of a delimiter-separated string}
\usage{
dsv_colnames(x, delim = ",", quote = "\\"")
}
\arguments{
\item{x}{Delimiter-separated string. A character scalar.}

\item{delim}{Single character used to separate fields within \code{x}.}

\item{quote}{Single character used to quote strings within \code{x}. Set to \code{NULL} for none.}
}
\value{
A character vector of column names.
}
\description{
Returns the column names of a string in a \href{https://en.wikipedia.org/wiki/Delimiter-separated_values}{delimiter-separated-value} format like
\href{https://en.wikipedia.org/wiki/Comma-separated_values}{CSV} or \href{https://en.wikipedia.org/wiki/Tab-separated_values}{TSV}.
}
\examples{
\donttest{
httr::GET("https://raw.githubusercontent.com/tidyverse/readr/master/inst/extdata/mtcars.csv") |>
  httr::content(as = "text") |>
  pal::dsv_colnames()}
}
\seealso{
Other string functions:
\code{\link{as_chr}()},
\code{\link{as_comment_string}()},
\code{\link{as_line_feed_chr}()},
\code{\link{as_string}()},
\code{\link{escape_lf}()},
\code{\link{fuse_regex}()},
\code{\link{prettify_nr}()},
\code{\link{prose_ls}()},
\code{\link{sentenceify}()},
\code{\link{wrap_chr}()}
}
\concept{string}
