% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{enum_fn_param_defaults}
\alias{enum_fn_param_defaults}
\title{Enumerate function's default parameter values}
\usage{
enum_fn_param_defaults(
  param,
  fn = sys.function(sys.parent()),
  env = parent.frame(),
  wrap = "`",
  sep2 = " or ",
  last = sep2,
  ...
)
}
\arguments{
\item{param}{Parameter name. A character scalar.}

\item{fn}{A \link[base:function]{function} or a function name (searched for in \code{env}). See \code{\link[base:formals]{base::formals()}} for details.}

\item{env}{\link[base:environment]{Environment} \code{fn} is defined in. See \code{\link[base:formals]{base::formals()}} for details.}

\item{wrap}{Character sequence the default parameter values are to be wrapped in. A character vector or something coercible to.}

\item{sep2, last}{Passed on to \code{\link[cli:ansi_collapse]{cli::ansi_collapse()}}.}

\item{...}{Further arguments passed on to \code{\link[cli:ansi_collapse]{cli::ansi_collapse()}}.}
}
\value{
A character scalar.
}
\description{
Convenience function combining \code{\link[=fn_param_defaults]{fn_param_defaults()}}, \code{\link[=wrap_chr]{wrap_chr()}} and \code{\link[cli:ansi_collapse]{cli::ansi_collapse()}}.
}
\details{
This function can be very convenient to avoid duplication in roxygen2 documentation by leveraging \href{https://roxygen2.r-lib.org/articles/rd-formatting.html#inline-code}{inline \R code evaluation} as follows:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{#' @param some_param Some parameter. One of
#'   `r pal::enum_fn_param_defaults(param = "some_param", fn = "some_fn")`.
#'
some_fn <- function(some_param = c("a", "b", "c")) \{
  some_param <- rlang::arg_match(some_param)
  ...
\}
}\if{html}{\out{</div>}}
}
\examples{
pal::enum_fn_param_defaults(param = ".name_repair",
                            fn = tibble::as_tibble) |>
  cat()
}
\seealso{
Other package documentation functions:
\code{\link{fn_param_defaults}()},
\code{\link{roxy_blocks}()},
\code{\link{roxy_obj}()},
\code{\link{roxy_tag_value}()},
\code{\link{roxy_to_md_links}()}
}
\concept{roxy}
