% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{escape_lf}
\alias{escape_lf}
\title{Escape line feeds / newlines}
\usage{
escape_lf(x, escape_cr = FALSE)
}
\arguments{
\item{x}{A character vector.}

\item{escape_cr}{Whether or not to also escape the carriage return character \code{CR} (aka \verb{\\r}). A logical scalar.}
}
\value{
A character vector of the same length as \code{x}.
}
\description{
Escapes the \href{https://en.wikipedia.org/wiki/Newline}{POSIX-standard newline control character \code{LF}} (aka \verb{\\n}) which is the standard on Unix/Linux and recent
versions of macOS. Set \code{escape_cr = TRUE} in order to also escape the carriage return character \code{CR} (aka \verb{\\r}) commonly used on Microsoft Windows.
}
\examples{
library(magrittr)

# read in and print RStudio add-in registration file as-is
text <-
  fs::path_package(package = "pal",
                   "rstudio", "addins.dcf") |>
  readr::read_file() \%T>\%
  pal::cat_lines()

# escape newlines and print again
pal::escape_lf(text) |> pal::cat_lines()
}
\seealso{
Other string functions:
\code{\link{as_chr}()},
\code{\link{as_comment_str}()},
\code{\link{as_line_feed_chr}()},
\code{\link{as_str}()},
\code{\link{capitalize_first}()},
\code{\link{dsv_colnames}()},
\code{\link{enum_str}()},
\code{\link{fuse_regex}()},
\code{\link{prettify_nr}()},
\code{\link{prose_ls}()},
\code{\link{sentenceify}()},
\code{\link{wrap_chr}()}
}
\concept{string}
