\name{exampleData}
\alias{exampleData}

\docType{data}
\title{Simulated expression data with 56 genes (rows) and 37 samples (cols)}
\description{Genes are named according to their Entrez ids and samples as Donor_Age. This toy data is generated by averaging randomly selected samples from the publicly available expression dataset E-MTAB-1724 (BabyDiet) from ArrayExpress}

\usage{exampleData}
\keyword{datasets}


