\name{exampleInfo}
\alias{exampleInfo}

\docType{data}
\title{Information about the artificial samples in exampleData}
\description{exampleInfo is a data frame with 37 rows (samples in exampleData) and 4 columns (variables). These variables are randomly generated, not measured from any real biological samples.}

\usage{exampleInfo}
\format{
	\describe{
		\item{Age}{Numeric values indicating the ages of the donors when the samples are extracted}
		\item{Donor}{Characters indicating the individuals from who the samples are extracted}
		\item{TimeToDiagnosis}{Numeric values describing how long after the sample extraction the donor is diagnosed}
		\item{SampleGroup}{Integers indicating sample groups (0 for controls, 1 for cases)}
	}
}
\keyword{datasets}


