% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{exists_in_namespace}
\alias{exists_in_namespace}
\title{Test if object exists in namespace}
\usage{
exists_in_namespace(x, ns, pos = -1L, envir = as.environment(pos))
}
\arguments{
\item{x}{an object name (given as a character string).}

\item{ns}{a namespace, or character string giving the namespace.}

\item{pos}{where to look for the object: see \code{\link{get}}.}

\item{envir}{an alternative way to specify an environment to look in.}
}
\value{
A logical scalar.
}
\description{
Determines whether or not an object exists in a specific \href{https://adv-r.hadley.nz/environments.html#namespaces}{namespace}.
}
\details{
This complements \code{\link[utils:getFromNamespace]{utils::getFromNamespace()}} and uses almost the same code internally.
}
\examples{
if (pal::is_pkg_installed("pkgpurl")) {
  pal::exists_in_namespace("pkg_config", "pkgpurl")
}

pal::exists_in_namespace("pkg_config", "pal")
}
