% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{fuse_regex}
\alias{fuse_regex}
\title{Fuse regular expressions}
\usage{
fuse_regex(..., .literal = FALSE)
}
\arguments{
\item{...}{Regular expressions. All elements will be converted to type character before fusing. \link[rlang:dyn-dots]{Dynamic dots} are supported.}

\item{.literal}{Whether or not the input should be interpreted \emph{literally} instead of as regular expressions. If \code{TRUE}, all elements of \code{...} are wrapped in
the literal escape sequences \verb{\\Q} and \verb{\\E}.}
}
\value{
A character scalar.
}
\description{
Combines a vector or list of regular expressions to a single one (by logical OR).
}
\examples{
\donttest{
# perform some (nonsense) Jane Austen text extraction
regex <- c("My dear Jane",
           "make haste, ",
           "(?i)\\\\bevil")

stringr::str_subset(string = janeaustenr::prideprejudice,
                    pattern = pal::fuse_regex(regex))}
}
\seealso{
The \link[rex:rex]{rex} package which provides an intuitive framework to build complex regular expressions.

Other string functions:
\code{\link{as_chr}()},
\code{\link{as_comment_str}()},
\code{\link{as_line_feed_chr}()},
\code{\link{as_str}()},
\code{\link{capitalize_first}()},
\code{\link{dsv_colnames}()},
\code{\link{enum_str}()},
\code{\link{escape_lf}()},
\code{\link{prettify_nr}()},
\code{\link{prose_ls}()},
\code{\link{sentenceify}()},
\code{\link{wrap_chr}()}
}
\concept{string}
