% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{gh_dir_ls}
\alias{gh_dir_ls}
\title{List files and directories in a GitHub repository}
\usage{
gh_dir_ls(
  path = "",
  owner,
  name,
  rev = "HEAD",
  recurse = TRUE,
  incl_dirs = TRUE,
  incl_files = TRUE
)
}
\arguments{
\item{path}{The path from the repository's root to the desired directory. A \link[fs:fs_path]{path} or character scalar.}

\item{owner}{The repository owner's GitHub user or organization name. A character scalar.}

\item{name}{The repository name. A character scalar.}

\item{rev}{The \href{https://git-scm.com/docs/revisions#Documentation/revisions.txt-emltrevgtltpathgtemegemHEADREADMEememmasterREADMEem}{Git revision expression}
matching the desired Git tree object, e.g. a branch name or another symbolic reference like \verb{HEAD@\{yesterday\}} or \code{HEAD~10}. A character scalar.}

\item{recurse}{Whether or not to recurse into subdirectories of \code{path}.}

\item{incl_dirs}{Whether or not to list directories (and subdirectories if \code{recurse = TRUE}).}

\item{incl_files}{Whether or not to list files (also inside subdirectories if \code{recurse = TRUE}).}
}
\value{
A character vector of paths from the repository root to the files and subdirectories found under \code{rev:path}.
}
\description{
List files and directories in a GitHub repository
}
\examples{
# you can opt-out from directory recursion
pal::gh_dir_ls(owner = "salim-b",
               name = "pal",
               recurse = FALSE) |>
  pal::cat_lines()

# or list only files or directories
pal::gh_dir_ls(path = "tests",
               owner = "salim-b",
               name = "pal",
               incl_files = FALSE) |>
  pal::cat_lines()

pal::gh_dir_ls(path = "tests",
               owner = "salim-b",
               name = "pal",
               incl_dirs = FALSE) |>
  pal::cat_lines()
}
\seealso{
Other GitHub functions:
\code{\link{gh_text_files}()},
\code{\link{gh_text_file}()}
}
\concept{gh}
