% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{gh_ls}
\alias{gh_ls}
\title{List files and directories in a GitHub repository}
\usage{
gh_ls(
  path,
  owner,
  name,
  branch = "main",
  recurse = TRUE,
  incl_dirs = TRUE,
  incl_files = TRUE
)
}
\arguments{
\item{path}{The path from the repository's root to the desired directory or file. A \link[fs:fs_path]{path} or character scalar.}

\item{owner}{The repository owner's GitHub user or organization name. A character scalar.}

\item{name}{The repository name. A character scalar.}

\item{branch}{The repository's desired branch name. A character scalar.}

\item{recurse}{Whether or not to recurse into subdirectories of \code{path}.}

\item{incl_dirs}{Whether or not to list directories (and subdirectories if \code{recurse = TRUE}).}

\item{incl_files}{Whether or not to list files (also inside subdirectories if \code{recurse = TRUE}).}
}
\value{
A character vector of paths from the repository root to the files and subdirectories found under \code{branch:path}.
}
\description{
List files and directories in a GitHub repository
}
\examples{
pal::gh_ls(path = "tests",
           owner = "salim-b",
           name = "pal",
           branch = "master")

# you can opt-out from directory recursion
pal::gh_ls(path = "tests",
           owner = "salim-b",
           name = "pal",
           branch = "master",
           recurse = FALSE)

# or list only files or directories
pal::gh_ls(path = "tests",
           owner = "salim-b",
           name = "pal",
           branch = "master",
           incl_files = FALSE)

pal::gh_ls(path = "tests",
           owner = "salim-b",
           name = "pal",
           branch = "master",
           incl_dirs = FALSE)
}
\seealso{
Other GitHub functions:
\code{\link{gh_text_files}()},
\code{\link{gh_text_file}()}
}
\concept{gh}
