% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{gh_text_file}
\alias{gh_text_file}
\title{Read in a text file from a GitHub repository}
\usage{
gh_text_file(path, owner, name, branch = "main")
}
\arguments{
\item{path}{The path from the repository's root to the desired directory or file. A \link[fs:fs_path]{path} or character scalar.}

\item{owner}{The repository owner's GitHub user or organization name. A character scalar.}

\item{name}{The repository name. A character scalar.}

\item{branch}{The repository's desired branch name. A character scalar.}
}
\value{
A character scalar, or an empty character vector in case no text file is found under \code{branch:path}.
}
\description{
Downloads the text file under the specified path from a GitHub repository via \href{https://docs.github.com/graphql}{GitHub's GraphQL API v4} and returns its
content as a string.
}
\details{
Works for both public and private repositories, for the latter you just need to set up a sufficiently authorized \link[gh:gh_token]{GitHub Personal Access Token (PAT)}.

Note that nothing is returned in case of a \href{https://en.wikipedia.org/wiki/Binary_file}{\emph{binary} file}, as if no file at all existed under the given
\code{path}.
}
\examples{
pal::gh_text_file(path = "pal.Rproj",
                  owner = "salim-b",
                  name = "pal",
                  branch = "master") |>
  cat()
}
\seealso{
Other GitHub functions:
\code{\link{gh_ls}()},
\code{\link{gh_text_files}()}
}
\concept{gh}
