% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{git_file_mod_time}
\alias{git_file_mod_time}
\title{Get Git file modification time}
\usage{
git_file_mod_time(path, repo = ".")
}
\arguments{
\item{path}{Path to a file in \code{repo}.}

\item{repo}{Path to a Git repository.}
}
\value{
A \link[base:DateTimeClasses]{datetime}.
}
\description{
Determine the time a file in a Git repository was last modified.
}
\details{
Note that only \emph{committed} changes to the file are regarded. The modification is returned in \href{https://en.wikipedia.org/wiki/Coordinated_Universal_Time}{UTC}.
}
\examples{
\dontrun{
pal::git_file_mod_time(path = "README.md")}
}
\seealso{
Other Git functions:
\code{\link{git_remote_tree_url}()}
}
\concept{git}
