% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{gitlab_document}
\alias{gitlab_document}
\title{Convert to GitLab Flavored Markdown}
\usage{
gitlab_document(smart_punctuation = TRUE, parse_emoji_markup = FALSE, ...)
}
\arguments{
\item{smart_punctuation}{Setting this to \code{TRUE} activates \href{https://pandoc.org/MANUAL.html#extension-smart}{Pandoc's \code{smart} extension} which converts
straight quotes to curly quotes, \verb{---} to an em-dash (—), \verb{--} to an en-dash (–), and \code{...} to ellipses (…). Nonbreaking spaces are inserted after certain
abbreviations, such as \code{Mr.}.}

\item{parse_emoji_markup}{Setting this to \code{TRUE} activates \href{https://pandoc.org/MANUAL.html#extension-emoji}{Pandoc's \code{emoji} extension} which parses emoji
markup (e.g. \verb{:smile:}) as Unicode emoticons.}

\item{...}{Additional arguments passed to \code{\link[rmarkdown:md_document]{rmarkdown::md_document()}}.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{rmarkdown::render()}}.
}
\description{
Format for converting from R Markdown to \href{https://gitlab.com/help/user/markdown.md}{GitLab Flavored Markdown}.
}
\details{
This is the GitLab equivalent to the \link[rmarkdown:github_document]{github_document} R Markdown
\href{https://bookdown.org/yihui/rmarkdown/output-formats.html}{output format}. It builds upon \link[rmarkdown:md_document]{md_document} and all of its parameters
can be specified except \code{variant}, \code{md_extensions} and \code{pandoc_args}.

This output format basically ensures Pandoc is called with a custom set of options optimized for maximum compatibility with
\href{https://gitlab.com/help/user/markdown.md}{GitLab Flavored Markdown}.
}
\examples{
\dontrun{
rmarkdown::render("input.Rmd", gitlab_document())

rmarkdown::render("input.Rmd", gitlab_document(parse_emoji_markup = TRUE,
                                               preserve_yaml = TRUE,
                                               extension = ".markdown"))
}
}
