% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{has_pkg_config_val}
\alias{has_pkg_config_val}
\title{Test if package configuration value is set}
\usage{
has_pkg_config_val(key, pkg)
}
\arguments{
\item{key}{Configuration key name. A character scalar.}

\item{pkg}{Package name. A character scalar.}
}
\value{
A logical scalar.
}
\description{
Tests whether or not a certain package configuration value is set. See \code{\link[=pkg_config_val]{pkg_config_val()}} for the underlying concept.
}
\details{
Note that \code{has_pkg_config_val()} throws an error if the package configuration \emph{key} doesn't exist.
}
\examples{
try(
  pal::has_pkg_config_val(key = "gen_pkgdown_ref",
                          pkg = "pkgpurl")
)
}
\seealso{
Other package configuration functions:
\code{\link{augment_pkg_config}()},
\code{\link{pkg_config_val_default}()},
\code{\link{pkg_config_val}()},
\code{\link{print_pkg_config}()}
}
\concept{pkg_config}
