% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{helper_options}
\alias{helper_options}
\alias{.chores_dir}
\alias{.chores_chat}
\title{Options used by the chores package}
\description{
The chores package makes use of three notable user-facing options:
\itemize{
\item \code{.chores_dir} is the directory where helper prompts live. See the helper \link{directory}
help-page for more information.
\item \code{.chores_chat} determines the underlying LLM powering each helper.
See the "Choosing a model" section of \code{vignette("chores", package = "chores")}
for more information.
}
}
