% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{is_installed}
\alias{is_installed}
\title{Test if packages are installed}
\usage{
is_installed(pkg)
}
\arguments{
\item{pkg}{Package names.}
}
\value{
A named logical vector of the same length as \code{pkg}.
}
\description{
This function returns \code{TRUE} for each \code{pkg} that is installed on the current system and \code{FALSE} otherwise.
}
\details{
In contrast to \code{\link[base:require]{base::require()}}, it checks if a package is installed without attaching its namespace if so.

In contrast to \code{\link[rlang:is_installed]{rlang::is_installed()}}, it is fully vectorized, i.e. returns a (named) logical vector of the same length as \code{pkg}.

It is
\href{https://stackoverflow.com/questions/9341635/check-for-installed-packages-before-running-install-packages/38082613#38082613}{considerably faster} than the
commonly used \code{pkg \%in\% rownames(installed.packages())} check.
}
\examples{
is_installed(pkg = "tidyverse")
}
\seealso{
R package functions
\code{\link{is_pkg_dir}()},
\code{\link{ls_pkg}()}
}
\concept{rpkgs}
