% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{is_pkg_cran}
\alias{is_pkg_cran}
\title{Test if a package is available on CRAN}
\usage{
is_pkg_cran(pkg, min_version = NULL)
}
\arguments{
\item{pkg}{Package name. A character scalar.}

\item{min_version}{Minimum required version number of \code{pkg}. Must be either \code{NULL} to ignore version numbers, or a single
\code{\link[base:numeric_version]{package_version}} or something coercible to.}
}
\value{
A logical scalar.
}
\description{
Test if a package is available on CRAN
}
\examples{
pal::is_pkg_cran("foobar")
pal::is_pkg_cran("dplyr")
pal::is_pkg_cran("dplyr", min_version = 9999.9)
}
