% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{is_pkgdown_dir}
\alias{is_pkgdown_dir}
\title{Test if pkgdown is set up for an R package directory}
\usage{
is_pkgdown_dir(path = ".")
}
\arguments{
\item{path}{Path of the R package directory to check. A character scalar. Defaults to the current working directory.}
}
\value{
\code{TRUE} if pkgdown is set up for \code{path}, \code{FALSE} otherwise.
}
\description{
Convenience wrapper around the \code{\link[rprojroot:criteria]{rprojroot::is_pkgdown_project}} root criterion. Note that it will only return \code{TRUE} for the
root of a package directory and the \code{pkgdown} subdirectory, not other subdirectories.
}
\examples{
pal::is_pkgdown_dir()
pal::is_pkgdown_dir(fs::path_package("pal"))
}
\seealso{
Other R package functions:
\code{\link{assert_pkg}()},
\code{\link{has_pkg_config_val}()},
\code{\link{is_pkg_cran}()},
\code{\link{is_pkg_dir}()},
\code{\link{is_pkg_installed}()},
\code{\link{ls_pkg}()},
\code{\link{pkg_config_val}()},
\code{\link{use_pkg}()}
}
\concept{rpkgs}
