% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{is_url}
\alias{is_url}
\title{Test if URL}
\usage{
is_url(x)
}
\arguments{
\item{x}{Character vector to test.}
}
\value{
A logical vector of the same length as \code{x}.
}
\description{
Tests whether the elements of a character vector are \href{https://de.wikipedia.org/wiki/Uniform_Resource_Locator}{Uniform Resource Locators} (URLs).
}
\details{
This function is based on \code{\link[xml2:url_parse]{xml2::url_parse()}} and simply checks whether the elements in \code{x} contain both a \strong{scheme} as well as \emph{some} \strong{scheme-specific
part} (excl. ports). No further checks are performed, so it's explicitly not verified that a URL actually conforms to its respective scheme.
}
\examples{
pal::is_url(c("/some/path",
              "file:///some/path"))
}
\seealso{
Other HTTP functions:
\code{\link{assert_mime_type}()},
\code{\link{is_http_success}()},
\code{\link{req_cached}()}
}
\concept{http}
