% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{knitr_table_format}
\alias{knitr_table_format}
\title{Determine current knitr table format}
\usage{
knitr_table_format(default = c("pipe", "simple", "html", "latex", "rst"))
}
\arguments{
\item{default}{knitr table format to fall back to when the \R option \code{knitr.table.format} is not set. One of
\itemize{
\item \code{"pipe"}
\item \code{"simple"}
\item \code{"html"}
\item \code{"latex"}
\item \code{"rst"}
}

See \code{\link[knitr:kable]{knitr::kable()}}'s \code{format} argument for details.}
}
\value{
A character scalar.
}
\description{
Determines the current knitr table format based on the \R option
\href{https://bookdown.org/yihui/rmarkdown-cookbook/kable.html#kable-formats}{\code{knitr.table.format}} which can either be set directly to a valid format string or
to a function returning one of these strings conditionally.
}
\details{
This is basically a convenience wrapper to be able to access the current \code{knitr.table.format} in a hassle-free way, i.e. it provides the conditional logic to
account for the possibility that \code{knitr.table.format} is set to a function rather than a format string.
}
\seealso{
Other R Markdown and knitr functions:
\code{\link{build_readme}()},
\code{\link{strip_yaml_header}()}
}
\concept{rmd_knitr}
