% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{list_drop}
\alias{list_drop}
\title{Drop list elements by name}
\usage{
list_drop(x, drop)
}
\arguments{
\item{x}{A list.}

\item{drop}{Names of the list elements to drop.}
}
\value{
A list.
}
\description{
Drops all list elements whose names match the provided set.
}
\examples{
as.list(mtcars) |> pal::list_drop(c("disp", "drat", "qsec"))

# works with any R objects that are internally of type list
mtcars |> pal::list_drop(c("disp", "drat", "qsec"))
}
\seealso{
Other list functions:
\code{\link{as_flat_list}()},
\code{\link{list_keep}()}
}
\concept{list}
