% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{list_keep}
\alias{list_keep}
\title{Keep list elements by name}
\usage{
list_keep(x, keep)
}
\arguments{
\item{x}{A list.}

\item{keep}{The names of the list elements to keep.}
}
\value{
A list.
}
\description{
Keeps only the list elements whose names match the provided set.
}
\examples{
as.list(mtcars) |> pal::list_keep(c("gear", "carb"))

# works with any R objects that are internally of type list
mtcars |> pal::list_keep(c("gear", "carb"))
}
\seealso{
Other list functions:
\code{\link{as_flat_list}()},
\code{\link{list_drop}()}
}
\concept{list}
