% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{md_xml}
\alias{md_xml}
\title{Parse (R) Markdown as CommonMark XML tree}
\usage{
md_xml(md, eol = c("LF", "CRLF", "CR", "LFCR"))
}
\arguments{
\item{md}{The (R) Markdown file as a character scalar.}

\item{eol}{End of line (EOL) control character sequence. One of
\itemize{
\item \code{"LF"} for the line feed (LF) character (\code{"\\n"}). The standard on Unix and Unix-like systems (Linux, macOS, *BSD, etc.) and the \strong{default}.
\item \code{"CRLF"} for the carriage return + line feed (CR+LF) character sequence (\code{"\\r\\n"}). The standard on Microsoft Windows, DOS and some other systems.
\item \code{"CR"} for the carriage return (CR) character (\code{"\\r"}). The standard on classic Mac OS and some other antiquated systems.
\item \code{"LFCR"} for the line feed + carriage return (LF+CR) character sequence (\code{"\\n\\r"}). The standard on RISC OS and some other exotic systems.
}}
}
\value{
An \code{\link[xml2:oldclass]{xml_node}} or \code{\link[xml2:oldclass]{xml_nodeset}} (possibly empty). Results are always de-duplicated.
}
\description{
Returns the CommonMark parse tree in XML format.
}
\examples{
pal::gh_text_file(path = "Rmd/pal.Rmd",
                  owner = "salim-b",
                  name = "pal") |>
pal::md_xml()
}
\concept{commonmark}
