% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{open_as_tmp_spreadsheet}
\alias{open_as_tmp_spreadsheet}
\title{Open as temporary spreadsheet}
\usage{
open_as_tmp_spreadsheet(x, format = c("csv", "xlsx"), quiet = TRUE)
}
\arguments{
\item{x}{A dataframe or \link[tibble:tibble]{tibble}, or something coercible to.}

\item{format}{The spreadsheet format to use. One of
\itemize{
\item \code{"csv"} for \href{https://en.wikipedia.org/wiki/Comma-separated_values}{comma-separated values} file written using \code{\link[readr:write_delim]{readr::write_csv()}}. The default.
\item \code{"xlsx"} for an \href{https://en.wikipedia.org/wiki/Office_Open_XML}{Office Open XML} file commonly used by Microsoft Excel 2007+, written using
\code{\link[writexl:write_xlsx]{writexl::write_xlsx()}}.
}}

\item{quiet}{Whether or not to echo the command to open the temporary spreadsheet on the console before running it.}
}
\value{
\code{x} invisibly.
}
\description{
This is a convenience function to write an R object -- usually tabular data like a dataframe or \link[tibble:tibble]{tibble} -- to a temporary spreadsheet and
and subsequently open that spreadsheet in the system's default application using \code{\link[xopen:xopen]{xopen::xopen()}}. The latter is usually equivalent to double-clicking on
the file in a file browser.
}
\examples{
\dontrun{
open_as_tmp_spreadsheet(mtcars,
                        format = "xslx")}
}
\seealso{
Data frame / Tibble functions
\code{\link{all_cols}()},
\code{\link{daff_diff}()},
\code{\link{is_equal_df}()}
}
\concept{tibble}
