% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{pal-class}
\alias{pal-class}
\title{An S4 class to represent pal attributes.}
\description{
An S4 class to represent pal attributes.
}
\section{Slots}{

\describe{
\item{\code{db}}{tbl_df object.}

\item{\code{global}}{.pal_global object.}

\item{\code{n}}{.pal_n object.}

\item{\code{colors}}{.pal_color object.}

\item{\code{tags}}{.pal_tags object.}
}}

