% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-package.R
\docType{package}
\name{pal-package}
\alias{pal}
\alias{pal-package}
\title{pal: Friendly Convenience/Utility Functions}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A collection of convenience/utility functions which aren't (yet) found in any other tidyverse-style R package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rpkg.dev/pal}
  \item \url{https://gitlab.com/salim_b/r/pkgs/pal}
  \item Report bugs at \url{https://gitlab.com/salim_b/r/pkgs/pal/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

}
\keyword{internal}
