% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-dir.R
\name{pal_add_dir}
\alias{pal_add_dir}
\alias{.pal_add_dir}
\alias{.pal_dir}
\alias{.pal_new}
\title{The pal directory}
\usage{
.pal_add_dir(dir)

.pal_dir()

.pal_new(role, interface)
}
\arguments{
\item{dir}{Path to a directory of markdown files--see \code{Details} for more.}

\item{role}{A single string giving a descriptor of the pal's functionality.}

\item{interface}{One of \code{"replace"}, \code{"prefix"}, or \code{"suffix"}, describing
how the pal will interact with the selection. For example, the
\link[=pal_cli]{cli pal} \code{"replace"}s the selection, while the
\link[=pal_roxygen]{roxygen pal} \code{"prefixes"} the selected code with documentation.}
}
\description{
Pals can be added in batch with \code{.pal_add_dir()}, which takes a directory
of markdown files. On package load, pal will add all pals defined in the pal
directory, located at \code{.pal_dir()} (by default, \verb{~/.config/pal}). To add a
new pal that will be loaded every time the package is loaded, use \code{.pal_new()}.
}
\details{
Prompts are markdown files with the
name \code{role-interface.md}, where interface is one of
"replace", "prefix" or "suffix".
An example directory might look like:

\if{html}{\out{<div class="sourceCode">}}\preformatted{/
├── .config/
│   └── pal/
│       ├── proofread-replace.md
│       └── summarize-prefix.md
}\if{html}{\out{</div>}}

In that case, pal will register two custom pals when you call \code{library(pal)}.
One of them has the role "proofread" and will replace the selected text with
a proofread version (according to the instructions contained in the markdown
file itself). The other has the role "summarize" and will prefix the selected
text with a summarized version (again, according to the markdown file's
instructions). Note:
\itemize{
\item Files without a \code{.md} extension are ignored.
\item Files with a \code{.md} extension must contain only one hyphen in their filename,
and the text following the hyphen must be one of \code{replace}, \code{prefix}, or
\code{suffix}.
}

To load custom prompts every time the package is loaded, place your
prompts in \verb{~/.config/pal} (or, to use some other folder, set
\code{options(.pal_dir = some_dir)} before loading the package).
}
