% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_chat.R
\name{pal_chat}
\alias{pal_chat}
\title{Apply pals to code}
\usage{
pal_chat(expr, pal = NULL)
}
\arguments{
\item{expr}{Lines of code that raise an error, as an expression.}

\item{pal}{A pal created with \code{\link[=pal]{pal()}}.}
}
\description{
Apply pals to code
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pal <- pal("cli")

pal_chat(stop("An error message."))
\dontshow{\}) # examplesIf}
}
