% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{pal_cnstr}
\alias{pal_cnstr}
\title{Create a pal object from a character vector or a list}
\usage{
pal_cnstr(col = NA)
}
\arguments{
\item{col}{numeric.}
}
\description{
Create a pal object from a character vector or a list
}
