% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\docType{methods}
\name{pal_colors}
\alias{pal_colors}
\title{Filter considering colors.}
\usage{
pal_colors(x, ..., dist = 75, only_alpha = TRUE)
}
\arguments{
\item{x}{A pal object.}

\item{\dots}{character, name of colors.}

\item{only_alpha}{logical}

\item{distance}{numeric, Distance to the colors.}
}
\description{
pal_colors allow filter palettes considering colors.
}
