% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R, R/pal_n.R
\docType{methods}
\name{pal_n}
\alias{pal_n}
\alias{pal_n,pal-method}
\alias{pal_n,}
\alias{pal}
\alias{missing,}
\title{Filter considering color complexity.}
\usage{
pal_n(x, min = 0L, max = Inf)

\S4method{pal_n}{pal}(x, min = 0L, max = Inf)
}
\arguments{
\item{x}{pal object;}

\item{min}{numeric; Minimum number of colors on the palette.}

\item{max}{numeric; Maximum number of colors on the palette.}
}
\description{
Palettes are constituted by a specific number of color.
pal_n allow filter considering the number of colors on the palette.
}
\examples{

pal_search() # Number of palettes : 77004
pal_search() + pal_n(min = 100, max = 120) # Number of palettes : 500

}
