% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R, R/pal_plot.R
\docType{methods}
\name{pal_plot}
\alias{pal_plot}
\alias{pal_plt}
\alias{pal_plot,character-method}
\alias{pal_plot,}
\alias{character}
\alias{pal_plot,list-method}
\alias{list}
\alias{pal_plot,pal-method}
\alias{pal}
\title{Display color palettes in a plot}
\usage{
\S4method{pal_plot}{character}(x, plotly = FALSE)

\S4method{pal_plot}{list}(x, plotly = FALSE)

\S4method{pal_plot}{pal}(x, plotly = FALSE)
}
\arguments{
\item{x}{character vector,list or pal object; if x is a
character vector or a list must be in hexadecimal notation
or a built-in color names which R knows about (see \code{colors()}).}

\item{plotly}{logical; TRUE/FALSE whether a plotly plot must be returned.}
}
\description{
pal_plot create an excellent visualization for
finding the color palette that best suit your
interests.
}
\examples{
#list
pal1 <- c("#2F3869","#A9ABC6","#ECFC96","#B8A75E")
pal2 <- c("#1E5CB4","#54C7CF","#FED71C","#A5262C")
palettes <- list(pal001 = pal1, pal002 = pal2)
pal_plot(palettes,plotly = TRUE)

#character
pal <- c('blue','red','yellow')
pal_plot(pal,plotly = FALSE)
}
