% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_plot.R
\name{pal_plot}
\alias{pal_plot}
\title{Display multiple palettes in a plot}
\usage{
pal_plot(colors, newpage = TRUE, plotly = FALSE)
}
\arguments{
\item{colors}{list or vector; colors to plots,
colors param must be in hexadecimal notation.}

\item{newpage}{logical; TRUE/FALSE whether open a new page.}

\item{plotly}{logical; TRUE/FALSE whether a plotly
visualization must be returned.}
}
\description{
pal_plot create an excellent visualization for
finding the color palette that best suit your
interests.
}
\examples{
pal1 <- c("#2F3869","#A9ABC6","#ECFC96","#B8A75E")
pal2 <- c("#1E5CB4","#54C7CF","#FED71C","#A5262C")
palettes <- list(pal1 = pal1, pal2 = pal2)
pal_plot(palettes,plotly = TRUE)
}
