% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-pal-rmd2qmd.R
\name{pal_rmd2qmd}
\alias{pal_rmd2qmd}
\title{The rmd2qmd pal}
\description{
Rmarkdown and Quarto are two implmentations for combining prose with
executable code to make additional content, such as words, sentances, figures,
tables, and so on.
}
\section{Cost}{


The system prompt from a pal includes something
}

\section{Gallery}{


As an example, the following set of bookdown options:

\if{html}{\out{<div class="sourceCode">}}\preformatted{```\{r srch, fig.width = 9, warning = FALSE, echo = FALSE, out.width = '100\%', fig.height = 4, fig.cap = "Some examples.",  message = FALSE\}
}\if{html}{\out{</div>}}

are converted to

\if{html}{\out{<div class="sourceCode">}}\preformatted{```\{r\}
#| label: srch
#| warning: false
#| echo: false
#| message: false
#| fig-width: 9
#| fig-height: 4
#| fig-cap: "Some examples."
#| out-width: 100\%
}\if{html}{\out{</div>}}

As of Novermber 2024, \code{\link[knitr:convert_chunk_header]{knitr::convert_chunk_header()}} converts it to

\if{html}{\out{<div class="sourceCode">}}\preformatted{```\{r\}
label = "srch",
fig.width = 9,
warning = FALSE,
echo = FALSE,
out.width = "100\%",
fig.height = 4,
fig.cap = "Some examples.",
message = FALSE
}\if{html}{\out{</div>}}
}

\section{Interfacing manually with the rmd2qmd pal}{


Pals are typically interfaced with via the pal addin. To call the rmd2qmd
pal directly, use:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{pal_rmd2qmd <- .init_pal("rmd2qmd")
}\if{html}{\out{</div>}}

Then, to submit a query, run:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{pal_rmd2qmd$chat(\{expr\})
}\if{html}{\out{</div>}}
}

