% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_search.R
\name{pal_search}
\alias{pal_search}
\title{Create a pal object}
\usage{
pal_search(catalogs = "cpt-city", type = c("continuous", "discrete"))
}
\arguments{
\item{catalogs}{character vector; Collections of palettes that
are used in the search. The palette collections available in pal package are: \cr
\itemize{
\item cpt-city: http://soliton.vm.bytemark.co.uk/pub/cpt-city/index.html \cr
\item color-hex: https://www.color-hex.com/color-palettes/ \cr
\item paletteer: https://github.com/EmilHvitfeldt/paletteer \cr
}}

\item{type}{character vector; Type of palette. The type: \cr
\itemize{
\item 'continuous' \cr
\item 'discrete' \cr
\item 'diverging' \cr
\item 'qualitative' \cr
\item 'sequential' \cr
}

The types: diverging, qualitative and sequential
only are supported by the 'paletteer'catalog.}
}
\value{
a pal object (see \linkS4class{pal}).
}
\description{
pal_search() initializes a S4 object. It is also
used to set global parameters.
}
