% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R, R/pal_colors.R, R/pal_tags.R
\docType{methods}
\name{pal_tags}
\alias{pal_tags}
\alias{pal_colors,pal-method}
\alias{pal_colors,}
\alias{pal}
\alias{pal_tags,pal-method}
\alias{pal_tags,}
\alias{pal_tags,character-method}
\alias{mn}
\title{Filter considering tags.}
\usage{
pal_tags(x, ...)

\S4method{pal_colors}{pal}(x, ..., dist = 75, only_alpha = TRUE)

\S4method{pal_tags}{pal}(x, ...)

\S4method{pal_tags}{character}(x, min, max)
}
\arguments{
\item{x}{A pal object.}

\item{\dots}{character, name of tags.}
}
\description{
pal_tags allow filter palettes considering the tags.
For the catalogs:
\href{https://www.color-hex.com/color-palettes/}{color-hex},
\href{http://soliton.vm.bytemark.co.uk/pub/cpt-city/}{cpt-city}
the tags were created scraping the website. For the
\href{https://emilhvitfeldt.github.io/paletteer/}{paletteer}
catalog the package and palette names were used by created the tags.
}
