% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{path_script}
\alias{path_script}
\title{Determine file path of executing script}
\usage{
path_script()
}
\value{
The file path to the executing script.
}
\description{
Determine file path of executing script
}
\seealso{
System interaction functions
\code{\link{check_cli}()}
}
\concept{sys}
