% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{pkg_config_val_default}
\alias{pkg_config_val_default}
\title{Get default package configuration value}
\usage{
pkg_config_val_default(key, pkg)
}
\arguments{
\item{key}{Configuration key name. A character scalar.}

\item{pkg}{Package name. A character scalar.}
}
\value{
\code{pkgsnip::return_lbl("r_obj")}
}
\description{
Retrieves a package configuration's default value from the package's configuration metadata (\verb{<pkg>::pkg_config$default_value}). If no default value is
specified (\code{NULL}), nothing is returned (\code{NULL}).
}
\examples{
try(
  pal::pkg_config_val_default(key = "gen_pkgdown_ref",
                              pkg = "pkgpurl")
)
}
\seealso{
Other package configuration functions:
\code{\link{augment_pkg_config}()},
\code{\link{has_pkg_config_val}()},
\code{\link{pkg_config_val}()},
\code{\link{print_pkg_config}()}
}
\concept{pkg_config}
