% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_pipe.R
\name{+.pal}
\alias{+.pal}
\title{Add components to a pal_search()}
\usage{
\method{+}{pal}(left, rigth)
}
\arguments{
\item{left}{A pal object.}

\item{rigth}{A pal_search method.}
}
\description{
Based on ggplot2 and magrittr sintaxis we use \code{+} or \code{\%>\%} for
constructing sophisticated searchings. It allows you
to start simple, then get more and more complex search,
checking your work at each step.
}
\examples{

pal_search() +
 pal_plot(return = 5, plotly = FALSE)

pal_search() \%>\%
 pal_plot(return = 5, plotly = FALSE)
}
