% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{prettify_nr}
\alias{prettify_nr}
\title{Prettify a numeric vector}
\usage{
prettify_nr(
  x,
  round_to = 0.1,
  round_up = TRUE,
  big_mark = "'",
  decimal_mark = ".",
  justify_right = FALSE,
  ...
)
}
\arguments{
\item{x}{A numeric vector to prettify.}

\item{round_to}{The number to round \code{x} to. A numeric scalar.}

\item{round_up}{Whether to round a remainder of exactly \code{to / 2} \emph{up} or not. Set to \code{FALSE} in order to round \emph{off}.}

\item{big_mark}{The character used between every 3 digits to separate thousands.}

\item{decimal_mark}{The character used to indicate the numeric decimal point. Only relevant if \code{x} does not solely consist of integers.}

\item{justify_right}{Whether to right-justify the results to a common width. See the \code{trim} parameter of \code{\link[base:format]{base::format()}} for details.}

\item{...}{Further arguments passed on to \code{\link[base:format]{base::format()}}.}
}
\value{
A character vector of the same length as \code{x}.
}
\description{
Prettifies a numeric vector by rounding, separating thousands and optionally other procedures. Basically a convenience wrapper around \code{\link[=round_to]{round_to()}} and
\code{\link[base:format]{format()}}.
}
\examples{
c(0.11, 11111.11) |> pal::prettify_nr()

c(0.11, 11111.11) |>
  pal::prettify_nr(justify_right = TRUE) |>
  pal::cat_lines()
}
\seealso{
Other string functions:
\code{\link{as_chr}()},
\code{\link{as_comment_string}()},
\code{\link{as_string}()},
\code{\link{cat_lines}()},
\code{\link{dsv_colnames}()},
\code{\link{escape_lf}()},
\code{\link{fuse_regex}()},
\code{\link{prose_ls}()},
\code{\link{sentenceify}()},
\code{\link{wrap_chr}()}
}
\concept{string}
