% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{print_pkg_config}
\alias{print_pkg_config}
\title{Print package configuration metadata}
\usage{
print_pkg_config(
  pkg = utils::packageName(env = parent.frame()),
  roxy_to_md = FALSE
)
}
\arguments{
\item{pkg}{Package name. A character scalar. Defaults to the name of the calling package.}

\item{roxy_to_md}{Whether or not to convert roxygen2 documentation \href{https://roxygen2.r-lib.org/articles/rd-formatting.html#function-links}{links in pseudo-Markdown style} to actual Markdown ones using \code{\link[=roxy_to_md_links]{roxy_to_md_links()}}.}
}
\value{
A character vector.
}
\description{
Prints a package's configuration metadata (\verb{<pkg>::pkg_config}) as a prettily formatted Markdown table.
}
\examples{
try(
  pal::print_pkg_config(pkg = "pkgpurl")
)
}
\seealso{
Other package configuration functions:
\code{\link{augment_pkg_config}()},
\code{\link{has_pkg_config_val}()},
\code{\link{pkg_config_val}()},
\code{\link{pkg_config_val_default}()}
}
\concept{pkg_config}
