% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{prose_ls}
\alias{prose_ls}
\title{List items concatenated in prose style (..., ... and ...)}
\usage{
prose_ls(x, wrap = "", sep = ", ", last_sep = " and ")
}
\arguments{
\item{x}{A vector or a list.}

\item{wrap}{String (usually a single character) in which \code{x} is to be wrapped.}

\item{sep}{Separator to delimit the elements of \code{x}.}

\item{last_sep}{Separator to delimit the second-last and last element of \code{x}.}
}
\value{
A character scalar.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \cr
This function has been deprecated in favor of the more powerful \code{\link[=enum_str]{enum_str()}} (which builds upon \code{\link[cli:ansi_collapse]{cli::ansi_collapse()}}) and will be removed in the future.
}
\details{
Takes a vector or list and concatenates its elements to a single string separated in prose-style.
}
\examples{
pal::prose_ls(1:5)
}
\seealso{
Other string functions:
\code{\link{as_chr}()},
\code{\link{as_comment_str}()},
\code{\link{as_line_feed_chr}()},
\code{\link{as_str}()},
\code{\link{capitalize_first}()},
\code{\link{dsv_colnames}()},
\code{\link{enum_str}()},
\code{\link{escape_lf}()},
\code{\link{fuse_regex}()},
\code{\link{prettify_nr}()},
\code{\link{sentenceify}()},
\code{\link{wrap_chr}()}
}
\concept{string}
