% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{reason_pkg_required}
\alias{reason_pkg_required}
\title{Assemble reason why package is required}
\usage{
reason_pkg_required(
  fn = rlang::call_name(rlang::caller_call()),
  pkg = utils::packageName(parent.frame())
)
}
\arguments{
\item{fn}{Name of the function that requires the \R package. A character scalar.}

\item{pkg}{Name of the package \code{fn} is part of. A character scalar.}
}
\value{
A character scalar.
}
\description{
Assembles a prose text containing the reason why a certain \R package is required by the \code{fn} of another \code{pkg}. Intended to be used as argument \code{reason} of
\code{\link[rlang:is_installed]{rlang::check_installed()}}.
}
\examples{
try(
  rlang::check_installed(pkg = "foo",
                         reason = pal::reason_pkg_required(pkg = "bar",
                                                           fn = "serve"))
)
}
