% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{reduce_df_list}
\alias{reduce_df_list}
\title{Reduce a nested list of data frames to a single data frame}
\usage{
reduce_df_list(x, strict = TRUE)
}
\arguments{
\item{x}{A list containing data frames at its leafs.}

\item{strict}{Ensure \code{x} contains data frames only and throw an error otherwise. If \code{FALSE}, leafs containing other objects are ignored (skipped).}
}
\value{
A data frame.
}
\description{
Recursively reduces a nested list containing data frames at its leafs to a single row-bound data frame.
}
\seealso{
Other data frame / tibble functions:
\code{\link{assert_cols}()},
\code{\link{is_equal_df}()}
}
\concept{df}
