% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{rename_from}
\alias{rename_from}
\title{Rename elements from dictionary}
\usage{
rename_from(x, dict, default = names(x))
}
\arguments{
\item{x}{Object whose elements are to be renamed.}

\item{dict}{A named character vector or list of strings where names are the old and the values are the new names.}

\item{default}{Value(s) used when names aren't matched by any name in \code{dict}. Recycled to the length of \code{x}.}
}
\value{
\code{x} with elements renamed according to \code{dict}.
}
\description{
Renames the elements of a vector or list from a dictionary, leaving unmatched names untouched by default.
}
\examples{
mtcars |> pal::rename_from(dict = c(mpg = "Miles/(US) gallon",
                                    cyl = "Number of cylinders",
                                    disp = "Displacement (cu.in.)",
                                    hp = "Gross horsepower",
                                    drat = "Rear axle ratio",
                                    wt = "Weight (1000 lbs)",
                                    qsec = "1/4 mile time",
                                    vs = "Engine (0 = V-shaped, 1 = straight)",
                                    am = "Transmission (0 = automatic, 1 = manual)",
                                    gear = "Number of forward gears",
                                    carb = "Number of carburetors",
                                    not_there = "Yikes!"))
}
