% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{rename_from_list}
\alias{rename_from_list}
\title{Rename elements from list}
\usage{
rename_from_list(x, names, default = names(x))
}
\arguments{
\item{x}{Object whose elements are to be renamed.}

\item{names}{A named list of strings where names are the old and the values are the new names.}

\item{default}{Value(s) used when names aren't matched by any name in \code{names}. Recycled to the length of \code{x}.}
}
\value{
\code{x} with elements renamed which occur in \code{names}.
}
\description{
Renames the elements of a vector or list from a names reference list, leaving unmatched names untouched by default.
}
\examples{
mtcars |> pal::rename_from_list(names = list(mpg = "Miles/(US) gallon",
                                             cyl = "Number of cylinders",
                                             disp = "Displacement (cu.in.)",
                                             hp = "Gross horsepower",
                                             drat = "Rear axle ratio",
                                             wt = "Weight (1000 lbs)",
                                             qsec = "1/4 mile time",
                                             vs = "Engine (0 = V-shaped, 1 = straight)",
                                             am = "Transmission (0 = automatic, 1 = manual)",
                                             gear = "Number of forward gears",
                                             carb = "Number of carburetors",
                                             not_there = "Yikes!"))
}
