% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{req_cached}
\alias{req_cached}
\title{Perform a cached HTTP GET request}
\usage{
req_cached(url, max_tries = 3L)
}
\arguments{
\item{url}{HTTP protocol address. The scheme is optional, so both \code{"google.com"} and \code{"https://google.com"} will work. A character scalar.}

\item{max_tries}{Maximum number of attempts to retry in case of an HTTP error. An integerish scalar.}
}
\value{
If request is successful (i.e. the request was successfully
performed and a response with HTTP status code <400 was recieved), an HTTP
\link[httr2]{response}; otherwise throws an error. Override this behaviour with
\code{\link[httr2:req_error]{req_error()}}.
}
\description{
Convenience wrapper around a bunch of \link[httr2:httr2-package]{httr2} functions.
}
\seealso{
Other HTTP functions:
\code{\link{assert_mime_type}()},
\code{\link{is_http_success}()},
\code{\link{is_url}()}
}
\concept{http}
