% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{round_to}
\alias{round_to}
\title{Round to any number}
\usage{
round_to(x, to = 0.2, round_up = TRUE)
}
\arguments{
\item{x}{A vector of numbers to round.}

\item{to}{The number to round \code{x} to. A numeric scalar.}

\item{round_up}{Whether to round a remainder of \code{to / 2} up or not.}
}
\value{
A numeric vector of the same length as \code{x}.
}
\description{
Round to any number
}
\examples{
c(0.1, 0.1999, 0.099999, 0.49, 0.55, 0.5, 0.9, 1) \%>\% round_to(0.05)
}
