% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{rowsum_gt_zero}
\alias{rowsum_gt_zero}
\title{Determine if row sum is greater than zero}
\usage{
rowsum_gt_zero(x)
}
\arguments{
\item{x}{An array of two or more dimensions, containing numeric, complex, integer or logical values, or a numeric data frame. Row sums are always calculated
over the first two dimensions, i.e. always the first two dimensions are regarded as rows.}
}
\value{
A logical vector. The \code{names} are taken from the original array.
}
\description{
This is a convenience function intended to be used in combination with dplyr's \code{\link[dplyr:across]{across()}} to turn it from its default "all of" into an
"any of" behavior. It's identical to the \code{rowAny()} helper function suggested in dplyr's
\href{https://dplyr.tidyverse.org/articles/colwise.html#how-do-you-convert-existing-code}{\code{"colwise"} vignette}.
}
\examples{
# Find all rows where *any* numeric variable is exactly 1
mtcars \%>\% dplyr::filter(rowsum_gt_zero(dplyr::across(where(is.numeric), ~ .x == 1L)))
}
