% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{rowwise_all}
\alias{rowwise_all}
\title{Apply \code{all()} rowwise to a data frame}
\usage{
rowwise_all(x)
}
\arguments{
\item{x}{A data frame or a tibble.}
}
\value{
A logical vector with the result of \code{all()} for each row of \code{x}.
}
\description{
This is a convenience function that helps to combine multiple \code{\link[dplyr:across]{across()}} statements in dplyr's \href{dplyr::filter()}{\code{filter()}} function.
See the examples.
}
\examples{
# select all rows where either `vs` and `am` is greater zero *or* `gear` and `carb` is greater two
mtcars \%>\% dplyr::filter(rowwise_all(dplyr::across(one_of("vs", "am"), ~ .x > 0))
                         | rowwise_all(dplyr::across(one_of("gear", "carb"), ~ .x > 2)))
}
