% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{roxy_blocks}
\alias{roxy_blocks}
\title{Get roxygen2 blocks}
\usage{
roxy_blocks(pkg = NULL, ..., text = NULL, quiet = TRUE)
}
\arguments{
\item{pkg, text}{Either a package name or a character vector of \R source code lines to extract the object's roxygen2 tag value from.}

\item{...}{Further arguments passed on to \code{\link[=download.packages]{download.packages()}}, excluding \code{pkgs}, \code{destdir}, \code{available} and \code{type}. Only relevant
if \code{pkg} is provided\link[rlang:dyn-dots]{Dynamic dots} are supported.)`}

\item{quiet}{Whether or not to suppress status output from internal processing.}
}
\value{
A list of \code{\link[roxygen2:roxy_block]{roxy_block}} objects.
}
\description{
Parses the roxygen2 package documentation of a specific R package or from a single \code{.R} source code file.
}
\examples{
# Either provide an R source file as a character vector `text` ...
text <- readr::read_lines(paste0("https://raw.githubusercontent.com/r-lib/rlang/",
                                 "db52a58d505b65f58ba922d4752b5b0061f2a98c/R/fn.R"))

pal::roxy_blocks(text = text) |> head(n = 3L)

# ... or provide a package name
\dontrun{
pal::roxy_blocks(pkg = "rlang",
                 repos = "https://cloud.r-project.org") |>
  head(n = 3L)}
}
\seealso{
Other package documentation functions:
\code{\link{roxy_obj}()},
\code{\link{roxy_tag_value}()}
}
\concept{roxy}
