% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{roxy_tag_value}
\alias{roxy_tag_value}
\title{Get an object's roxygen2 tag value(s)}
\usage{
roxy_tag_value(blocks, obj_name, tag_names = "param", param_name)
}
\arguments{
\item{blocks}{A list of \code{\link[roxygen2:roxy_block]{roxy_block}} objects as returned by \code{\link[=roxy_blocks]{roxy_blocks()}}.}

\item{obj_name}{Object name to extract, usually a function name. A character scalar.}

\item{tag_names}{Name(s) of the \href{https://roxygen2.r-lib.org/articles/rd.html}{roxygen2 tags} (without the \code{@}) to extract the value(s) from. A character
vector.}

\item{param_name}{Parameter name to extract the value from. Only relevant if \code{"param" \%in\% tag_names}. A character scalar.}
}
\value{
A character vector of the same length as \code{tag_names}.
}
\description{
Extracts the value(s) belonging to the \code{tag_name}s documenting \code{obj_name} from a list of \code{\link[roxygen2:roxy_block]{roxy_block}} objects.
}
\examples{
\dontrun{
blocks <- pal::roxy_blocks(pkg = "dplyr")

pal::roxy_tag_value(blocks = blocks,
                    obj_name = "across",
                    tag_names = "param",
                    param_name = ".fns") |>
  cat()}
}
\seealso{
Other package documentation functions:
\code{\link{roxy_blocks}()},
\code{\link{roxy_obj}()}
}
\concept{roxy}
