% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{safe_max}
\alias{safe_max}
\title{Maximum (safe)}
\usage{
safe_max(..., rm_na = TRUE)
}
\arguments{
\item{...}{Numeric objects of which to determine the maximum. \link[rlang:dyn-dots]{Dynamic dots} are supported.}

\item{rm_na}{Ignore missing values in \code{...}. If missing values are present and \code{rm_na = FALSE}, the result will always be \code{NA}.}
}
\value{
A numeric scalar or empty value, of the same type as \code{...}.
}
\description{
Modified version of \code{\link[base:Extremes]{base::max()}} that differs in the following ways:
\itemize{
\item \code{NA}s in the input (\code{...}) are ignored \emph{by default} (\code{rm_na = TRUE}).
\item If the input is of length zero, the output will also be of length zero (of the same type as the input).
\item It is ensured that all inputs are either numeric, of length zero or \code{NA}. The only case where the return value will be \code{-Inf} or \code{NA} is when the input
contains only \code{-Inf} or \code{NA}.
\item \link[rlang:dyn-dots]{Dynamic dots} are supported.
}
}
\examples{
# other than `base::max()`, this function removes `NA`s by default
max(1, NA_real_, 2)
pal::safe_max(1, NA_real_, 2)

# other than `base::max()`, this function does not return `-Inf` or `NA_character_` for
# zero-length inputs
max(NULL)
max(integer())
pal::safe_max(NULL)
pal::safe_max(integer())

# other than `base::max()`, this function fails for non-numeric inputs
max("zero", 1L)
max("zero", "one")
max(character())
try(pal::safe_max("zero", 1L))
try(pal::safe_max("zero", "one"))
try(pal::safe_max(character()))
}
\seealso{
Other statistical computing / numeric functions:
\code{\link{round_to}()},
\code{\link{safe_min}()},
\code{\link{safe_seq_len}()},
\code{\link{stat_mode}()}
}
\concept{stat}
