% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{script_path}
\alias{script_path}
\title{Determine file path of executing script}
\usage{
script_path()
}
\value{
The file path to the executing script.
}
\description{
Determine file path of executing script
}
