% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{sentenceify}
\alias{sentenceify}
\title{Convert to sentence case with trailing punctuation mark}
\usage{
sentenceify(x, punctuation_mark = ".")
}
\arguments{
\item{x}{Input to be converted to \href{https://en.wikipedia.org/wiki/Letter_case#Case_styles}{sentence case}, typically a character vector.}

\item{punctuation_mark}{Punctuation mark to be appended to \code{x}. A character scalar.}
}
\value{
A character vector of the same length as \code{x}.
}
\description{
Converts the input to a character vector and ensures it starts with an upper case letter and ends with the specified punctuation mark.
}
\details{
Note that this function doesn't alter any characters in \code{x} other than the first and the last one.
}
\examples{
pal::sentenceify("no verb, no sentence")

# punctuation mark won't be duplicated if already existing
pal::sentenceify(c("I've made my point",
                   "good point."))
}
\seealso{
\code{\link[stringr:case]{stringr::str_to_sentence()}} to convert a character vector to all lowercase except for the first character. Note that this also includes lowercasing \href{https://en.wikipedia.org/wiki/Proper_and_common_nouns}{proper nouns}, \href{https://en.wikipedia.org/wiki/Abbreviation}{abbreviations} etc.

\code{\link[snakecase:caseconverter]{snakecase::to_sentence_case()}} that builds upon \code{\link[stringr:case]{stringr::str_to_sentence()}} but offers additional options to finetune the conversion. Note that
\code{abbreviations} have to be manually specified in order to be preserved in upper case.

Other string functions:
\code{\link{as_chr}()},
\code{\link{as_comment_str}()},
\code{\link{as_env_var_name}()},
\code{\link{as_line_feed_chr}()},
\code{\link{as_str}()},
\code{\link{capitalize_first}()},
\code{\link{dsv_colnames}()},
\code{\link{enum_str}()},
\code{\link{escape_lf}()},
\code{\link{fuse_regex}()},
\code{\link{prettify_nr}()},
\code{\link{wrap_chr}()}
}
\concept{string}
