% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{set_attribute}
\alias{set_attribute}
\title{Set an attribute}
\usage{
set_attribute(object, attribute, value)
}
\arguments{
\item{object}{The object for which an attribute is to be changed.}

\item{attribute}{The attribute to be changed. A character scalar.}

\item{value}{The new value for the attribute.}
}
\value{
The \code{object} with the updated attribute, invisibly.
}
\description{
This is simply a \href{https://magrittr.tidyverse.org/articles/magrittr.html}{pipeable} version of \code{\link[base:attr]{attr(x, which) <- value}}. Unfortunately,
\code{\link[rlang:set_attrs]{rlang::set_attrs()}}, a more powerful attribute setter, has been marked deprecated.
}
\examples{
library(magrittr)

list(some = "element") \%>\%
  set_attribute(attribute = "custom_attribute",
                value = "foo") \%>\%
  print()
}
