% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{stat_mode}
\alias{stat_mode}
\title{Statistical mode}
\usage{
stat_mode(x, type = c("one", "all", "n"), na_rm = FALSE)
}
\arguments{
\item{x}{An \R object.}

\item{type}{What the function should calculate.
\itemize{
\item \code{"one"}: Return \emph{the} mode of \code{x}. If multiple modes or no mode at all exists, \code{NA} is returned.
\item \code{"all"}: Return \emph{all} modes of \code{x}. If none exists (e.g. because all values of \code{x} are distinct), \code{NA} is returned.
\item \code{"n"}: Return the number of modes of \code{x}.
}}

\item{na_rm}{Ignore missing values in \code{x}. A logical scalar.}
}
\value{
If \code{type = "n"}, the number of modes in \code{x} (an integer). Otherwise, the mode(s) of \code{x} or \code{NA} if none exist(s) (same type as \code{x}).
}
\description{
Compute the \href{https://en.wikipedia.org/wiki/Mode_(statistics)}{statistical mode} of a set of values. The mode is defined as the most frequent value, i.e. the
value that is most likely to be sampled.
}
\details{
See the package \href{https://cran.r-project.org/package=modeest}{modeest} for more powerful mode estimation functions.
}
\examples{
stat_mode(c(rep(3L, times = 3), 1:9))
stat_mode(c(1.5, 4, 9.9))

# if no mode exists, `NA` (of the same type as x) is returned
stat_mode(letters)
stat_mode(c(letters, "a"))

# if multiple modes exist, `NA` is returned by default
stat_mode(c(letters, "a", "b"))
# set `type = "all"` to return all modes instead
stat_mode(c(letters, "a", "b"),
          type = "all")

# `NA` is treated as any other value by default
stat_mode(c(letters, "a", NA_character_, NA_character_),
          type = "all")
# set `rm_na = TRUE` to ignore `NA` values
stat_mode(c(letters, "a", NA_character_, NA_character_),
          type = "all",
          na_rm = TRUE)
}
