% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{str_replace_verbose}
\alias{str_replace_verbose}
\title{Replace matched patterns in a string \emph{verbosely}}
\usage{
str_replace_verbose(string, pattern, verbose = TRUE)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{pattern}{A named character vector with patterns as names and replacements as values (\code{c(pattern1 = replacement1)}). Patterns are interpreted as
regular expressions as described in \code{\link[stringi:about_search_regex]{stringi::stringi-search-regex()}}. Replacements are interpreted as-is, except that references of the form \verb{\\1}, \verb{\\2},
etc. will be replaced with the contents of the respective matched group (created in patterns using \verb{()}). Pattern-replacement pairs are processed in the
order given, meaning that first listed pairs are applied before later listed ones.}

\item{verbose}{Whether to display replacements on the console.}
}
\value{
A character vector.
}
\description{
Apply a series of regular-expression-replacement pairs to a string. All performed replacements are displayed on the console by default.
}
\details{
This function provides a subset of \code{\link[stringr:str_replace]{stringr::str_replace_all()}}'s functionality. If you don't need the visual console output, it's recommended to directly
resort to that function.
}
\examples{
"Make love, not war" \%>\% str_replace_verbose(pattern = c("love" = "hummus",
                                                         "war" = "walls"))
}
\seealso{
String functions
\code{\link{as_chr}()},
\code{\link{as_string}()},
\code{\link{cat_lines}()},
\code{\link{dsv_colnames}()},
\code{\link{escape_lf}()},
\code{\link{fuse_regex}()},
\code{\link{prose_ls}()},
\code{\link{str_replace_file}()}
}
\concept{string}
