% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{strip_md}
\alias{strip_md}
\title{Strip Markdown formatting from character vector}
\usage{
strip_md(x, strip_footnotes = TRUE)
}
\arguments{
\item{x}{A character vector to strip Markdown formatting from.}

\item{strip_footnotes}{Whether to remove Markdown footnotes, too.}
}
\value{
A character vector of the same length as \code{x}.
}
\description{
Removes all Markdown formatting from a character vector.
}
\details{
This function relies on \code{\link[commonmark:commonmark]{commonmark::markdown_text()}} which \href{https://github.com/jeroen/commonmark#readme}{supports the CommonMark specification plus the Github extensions}. Unfortunately, \href{https://pandoc.org/MANUAL.html#footnotes}{Markdown footnotes} aren't supported
(yet). Therefore a separate option \code{strip_footnotes} is offered which relies on a simple regular expression to remove inline footnotes and footnote
references.
}
\examples{
strip_md("A **MD** formatted [string](https://en.wikipedia.org/wiki/String_(computer_science))")
}
