% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{strip_md_footnotes}
\alias{strip_md_footnotes}
\title{Strip Markdown footnotes from character vector}
\usage{
strip_md_footnotes(x)
}
\arguments{
\item{x}{A character vector to strip Markdown footnotes from. Note that elements in \code{x} are processed as separate Markdown domains, i.e. \emph{not} as individual
lines belonging to the same Markdown document.}
}
\value{
A character vector of the same length as \code{x}.
}
\description{
Removes all Markdown footnotes from a character vector.
}
\details{
Note that it is not checked whether footnote references and definitions actually match (by identifier), thus they are removed even if they were invalid.
}
\examples{
pal::strip_md_footnotes(
  "An **MD** formatted string with footnote[^fn].\n\n[^fn]: A note.\n"
)
}
\seealso{
Other (Pandoc) Markdown functions:
\code{\link{as_md_list}()},
\code{\link{as_md_val_list}()},
\code{\link{as_md_vals}()},
\code{\link{md_verb}()},
\code{\link{pipe_table}()},
\code{\link{strip_md}()}
}
\concept{md}
