% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{strip_yaml_header}
\alias{strip_yaml_header}
\title{Strip YAML header from R Markdown}
\usage{
strip_yaml_header(rmd, eol = c("LF", "CRLF", "CR", "LFCR"))
}
\arguments{
\item{rmd}{R Markdown file content as a character scalar.}

\item{eol}{End of line (EOL) control character sequence. One of
\itemize{
\item \code{"LF"} for the line feed (LF) character (\code{"\\n"}). The standard on Unix and Unix-like systems (Linux, macOS, *BSD, etc.) and the \strong{default}.
\item \code{"CRLF"} for the carriage return + line feed (CR+LF) character sequence (\code{"\\r\\n"}). The standard on Microsoft Windows, DOS and some other systems.
\item \code{"CR"} for the carriage return (CR) character (\code{"\\r"}). The standard on classic Mac OS and some other antiquated systems.
\item \code{"LFCR"} for the line feed + carriage return (LF+CR) character sequence (\code{"\\n\\r"}). The standard on RISC OS and some other exotic systems.
}}
}
\value{
Body of the R Markdown file as a character vector of lines.
}
\description{
Extracts the body from R Markdown file content, stripping a possible \href{https://bookdown.org/yihui/rmarkdown-cookbook/rmarkdown-anatomy.html#yaml-metadata}{YAML metadata block} at the beginning.
}
\details{
Note that for the \href{https://rmarkdown.rstudio.com/}{R Markdown file format}, the \href{https://pandoc.org/MANUAL.html#extension-yaml_metadata_block}{YAML metadata block} must occur at the beginning of the document (and there can be only one). Additional
whitespace characters (incl. newlines) before the YAML metadata block are allowed.
}
\examples{
rmd <- pal::gh_text_file(path = "README.Rmd",
                         owner = "salim-b",
                         name = "pal")
rmd |>
  stringr::str_split_1("\n") |>
  head() |>
  pal::cat_lines()

rmd |>
  pal::strip_yaml_header() |>
  head() |>
  pal::cat_lines()
}
\seealso{
Other R Markdown and knitr functions:
\code{\link{build_readme}()},
\code{\link{knitr_table_format}()}
}
\concept{rmd_knitr}
