% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{toml_read}
\alias{toml_read}
\title{Read in and parse TOML file as strict list}
\usage{
toml_read(input, from_file = TRUE, verbose = FALSE)
}
\arguments{
\item{input}{If \code{from_file = TRUE}, the path to a TOML file as a character scalar. Otherwise, TOML content as a character vector.}

\item{from_file}{Whether \code{input} is the path to a TOML file or already a character vector of TOML content.}

\item{verbose}{[logical] Optional verbosity flag, no current effect}
}
\value{
A \link[xfun:strict_list]{strict list}.
}
\description{
Reads in a file in \href{https://toml.io/}{Tom's Obvious Minimal Language (TOML)} format and returns its content as a (nested) \link[xfun:strict_list]{strict list}.
}
\details{
The file is parsed using \code{\link[RcppTOML:parseTOML]{RcppTOML::parseTOML(escape = FALSE)}}.
}
\seealso{
Other TOML functions:
\code{\link{toml_validate}()}
}
\concept{toml}
