% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{use_pkg}
\alias{use_pkg}
\title{Depend on another package}
\usage{
use_pkg(package, type = "Imports", min_version = TRUE, tidy = TRUE)
}
\arguments{
\item{package}{Name of package to depend on.}

\item{type}{Type of dependency: must be one of "Imports", "Depends",
"Suggests", "Enhances", or "LinkingTo" (or unique abbreviation). Matching
is case insensitive.}

\item{min_version}{Optionally, supply a minimum version for the package.
Set to \code{TRUE} to use the currently installed version.}

\item{tidy}{Whether or not to run \code{\link[usethis:tidyverse]{usethis::use_tidy_description()}} after adding \code{package} to dependencies.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Wrapper around \code{\link[usethis:use_package]{usethis::use_package()}} with different defaults and optional \code{DESCRIPTION} file \link[usethis:tidyverse]{tidying}.
}
\seealso{
Other R package functions:
\code{\link{assert_pkg}()},
\code{\link{augment_pkg_config}()},
\code{\link{has_pkg_config_val}()},
\code{\link{is_pkg_cran}()},
\code{\link{is_pkg_dir}()},
\code{\link{is_pkg_installed}()},
\code{\link{is_pkgdown_dir}()},
\code{\link{ls_pkg}()},
\code{\link{pkg_config_val}()}
}
\concept{rpkgs}
