% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{wrap_chr}
\alias{wrap_chr}
\title{Wrap character vector in string}
\usage{
wrap_chr(x, wrap = "\\"")
}
\arguments{
\item{x}{A character vector or something coercible to. Will be fed to \code{\link[=as_chr]{as_chr()}} before wrapping.}

\item{wrap}{Character sequence \code{x} is to be wrapped in. A character vector or something coercible to.}
}
\value{
A character vector of the same length as \code{wrap}.
}
\description{
Wrap character vector in string
}
\examples{
library(magrittr)

mtcars \%>\%
  magrittr::set_colnames(pal::wrap_chr(x = colnames(.),
                                       wrap = "`")) \%>\%
  pal::pipe_table()
}
\seealso{
Other string functions:
\code{\link{as_chr}()},
\code{\link{as_comment_string}()},
\code{\link{as_line_feed_chr}()},
\code{\link{as_string}()},
\code{\link{cat_lines}()},
\code{\link{dsv_colnames}()},
\code{\link{escape_lf}()},
\code{\link{fuse_regex}()},
\code{\link{prettify_nr}()},
\code{\link{prose_ls}()},
\code{\link{sentenceify}()}
}
\concept{string}
