% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{write_widget}
\alias{write_widget}
\title{Write HTML widget to file}
\usage{
write_widget(x, path)
}
\arguments{
\item{x}{HTML widget object.}

\item{path}{File path to write the HTML widget's core tags to.}
}
\value{
The content written to \code{path} as a character scalar, invisibly.
}
\description{
Writes an HTML widget's main container \verb{<div>} and data \verb{<script>} tags to file.
}
\details{
Other than \code{\link[htmlwidgets:saveWidget]{htmlwidgets::saveWidget()}}, this function does \emph{not} write a complete HTML document but only the widget's core tags. Hence it is intended to
create files which are in turn included in another HTML file, e.g. via a static site generator.

The HTML widget's dependencies must be written to files separately, e.g. using \code{\link[=write_widget_deps]{write_widget_deps()}}.
}
\seealso{
Other HTML widget functions:
\code{\link{write_widget_deps}()}
}
\concept{htmlwidget}
