% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{write_widget_deps}
\alias{write_widget_deps}
\title{Write HTML widget dependencies}
\usage{
write_widget_deps(
  x,
  path,
  path_base,
  path_head_snippet = NULL,
  quiet = !is.null(path_head_snippet)
)
}
\arguments{
\item{x}{HTML widget object.}

\item{path}{Directory path to write the HTML widget's dependencies to. A subfolder will be created for each dependency.}

\item{path_base}{Base path to determine the relative root of \code{path} for inclusion of the widget's dependencies. Must be a parent of or equal to \code{path}.}

\item{path_head_snippet}{Optional file path to write the snippet for inclusion of the widget's dependencies in the \verb{<head>} of an HTML document to.}

\item{quiet}{Whether or not to print the snippet for inclusion of the widget's dependencies to console.}
}
\value{
An \link[htmltools:HTML]{HTML} object suitable for inclusion in the \verb{<head>} of an HTML document, invisibly.
}
\description{
Writes an HTML widget's JS and CSS dependencies to files and returns a snippet for their inclusion in the \verb{<head>} of an HTML document.
}
\details{
To write the HTML widget's core tags to file, use \code{\link[=write_widget]{write_widget()}}.
}
\seealso{
Other HTML widget functions:
\code{\link{write_widget}()}
}
\concept{htmlwidget}
