% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{xml_to_md}
\alias{xml_to_md}
\title{Convert from CommonMark XML to (R) Markdown}
\usage{
xml_to_md(xml)
}
\arguments{
\item{xml}{CommonMark parse tree. An \code{\link[xml2:oldclass]{xml_document}}, \code{\link[xml2:oldclass]{xml_nodeset}} or
\code{\link[xml2:oldclass]{xml_node}}.}
}
\value{
A character scalar.
}
\description{
Convert from CommonMark XML to (R) Markdown
}
\examples{
pal::gh_text_file(path = "Rmd/pal.Rmd",
                  owner = "salim-b",
                  name = "pal") |>
  pal::md_to_xml() |>
  xml2::xml_contents() |>
  magrittr::extract(22:28) |>
  pal::xml_to_md() |>
  cat()
}
\seealso{
Other CommonMark parsing functions:
\code{\link{md_to_xml}()},
\code{\link{md_xml_subnode_ix}()}
}
\concept{commonmark}
