% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{as_chr}
\alias{as_chr}
\title{Convert to a character vector}
\usage{
as_chr(..., .use_names = FALSE)
}
\arguments{
\item{...}{\R objects to be converted to a character vector. \link[rlang:dyn-dots]{Dynamic dots} are supported.}

\item{.use_names}{Whether or not to preserve names by \code{\link[=unlist]{unlist()}}.}
}
\value{
A character vector.
}
\description{
\emph{Recursively} applies \code{\link[=as.character]{as.character()}} to its inputs.
}
\examples{
library(magrittr)

to_convert <-
  list(tibble::tibble(a = 1:3), "A", factor("wonderful"), xfun::strict_list("day")) \%T>\%
  print()

as.character(to_convert)
pal::as_chr(!!!to_convert)
}
\seealso{
Other string functions:
\code{\link{as_comment_string}()},
\code{\link{as_line_feed_chr}()},
\code{\link{as_string}()},
\code{\link{capitalize_first}()},
\code{\link{dsv_colnames}()},
\code{\link{escape_lf}()},
\code{\link{fuse_regex}()},
\code{\link{prettify_nr}()},
\code{\link{prose_ls}()},
\code{\link{sentenceify}()},
\code{\link{wrap_chr}()}
}
\concept{string}
