% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{assert_pkg}
\alias{assert_pkg}
\title{Assert a package is installed}
\usage{
assert_pkg(pkg, message = NULL)
}
\arguments{
\item{pkg}{Package name. A character scalar.}

\item{message}{The error message to display in case the package is not installed. If \code{NULL}, defaults to a sensible standard message.}
}
\value{
The package name invisibly.
}
\description{
Assert a package is installed
}
\examples{
assert_pkg("pal")

assert_pkg(pkg = "glue",
           message = paste0("You should really consider to install the awesome `glue` package! ",
                            "It's the glue that keeps strings and variables together 🤲."))
}
