% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{assert_pkg}
\alias{assert_pkg}
\title{Assert a package is installed}
\usage{
assert_pkg(pkg, min_version = NULL, message = NULL, install_hint = NULL)
}
\arguments{
\item{pkg}{Package name. A character scalar.}

\item{min_version}{Minimum required version number of \code{pkg}. Must be either \code{NULL} to ignore version numbers, or a single
\code{\link[base:numeric_version]{package_version}} or something coercible to.}

\item{message}{The error message to display in case the package is not installed. If \code{NULL}, a sensible standard message is generated.}

\item{install_hint}{Additional package installation instructions appended to \code{message}. Either \code{NULL} in order to autogenerate the hint, or a
character scalar. Set \code{install_hint = ""} in order to disable the hint.}
}
\value{
\code{pkg} invisibly.
}
\description{
Assert a package is installed
}
\examples{
assert_pkg("pal")

assert_pkg(pkg = "glue",
           message = paste0("You should really consider to install the awesome `glue` package! ",
                            "It's the glue that keeps strings and variables together 🤲."))
\dontrun{
assert_pkg("pal", min_version = 9999.9)

assert_pkg("yay",
           install_hint = paste0("To install the latest development version, run ",
                                 "`remotes::install_gitlab(\"salim_b/r/pkgs/yay\")`."))}
}
\seealso{
Other R package functions:
\code{\link{desc_list}()},
\code{\link{desc_value}()},
\code{\link{is_pkg_dir}()},
\code{\link{is_pkg_installed}()},
\code{\link{is_pkgdown_dir}()},
\code{\link{ls_pkg}()},
\code{\link{prose_ls_fn_param}()}
}
\concept{rpkgs}
