% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{augment_pkg_config}
\alias{augment_pkg_config}
\title{Augment package configuration metadata}
\usage{
augment_pkg_config(pkg)
}
\arguments{
\item{pkg}{Package name. A character scalar.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with at minimum the columns \code{key}, \code{default_value}, \code{r_opt} and \code{env_var}.
}
\description{
Augments a package's configuration metadata (\verb{<pkg>::pkg_config}) with the columns \code{r_opts} and \code{env_var} holding the respective \R option and environment
variable names.
}
\examples{
try(
  pal::augment_pkg_config(pkg = "pkgpurl")
)
}
\seealso{
Other package configuration functions:
\code{\link{has_pkg_config_val}()},
\code{\link{pkg_config_val_default}()},
\code{\link{pkg_config_val}()},
\code{\link{print_pkg_config}()}
}
\concept{pkg_config}
