% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{build_readme}
\alias{build_readme}
\title{Build \code{README.Rmd}}
\usage{
build_readme(
  input = "README.Rmd",
  output = "README.md",
  build_index_md = NULL,
  env = parent.frame(),
  clean = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{input}{Path to the R Markdown README file to be built. A character scalar.}

\item{output}{Path to the built Markdown README. A character scalar.}

\item{build_index_md}{Whether to build a separate \link[pkgdown:pkgdown-package]{pkgdown}-optimized \code{pkgdown/index.md} alongside \code{output} (i.e. in the same parent
directory). If \code{NULL}, it will only be built if the parent directory of \code{output} \link[=is_pkgdown_dir]{contains a pkgdown configuration file}. Note that it will
be built with the \R option \code{pal.build_readme.is_pkgdown} set to \code{TRUE}, allowing for conditional content inclusion in \code{input} – e.g. via the \href{https://yihui.org/knitr/options/#code-evaluation}{code chunk option} \code{eval = isTRUE(getOption("pal.build_readme.is_pkgdown"))}.}

\item{env}{Environment in which code chunks are to be evaluated, e.g. \code{\link[base:sys.parent]{base::parent.frame()}}, \code{\link[base:environment]{base::new.env()}}, or \code{\link[base:environment]{base::globalenv()}}.}

\item{clean}{Whether or not to delete intermediate files created by \code{\link[rmarkdown:render]{rmarkdown::render()}}.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
The path to \code{input} as a character scalar, invisibly.
}
\description{
A simpler, but considerably faster alternative to \code{\link[devtools:build_rmd]{devtools::build_readme()}} since it doesn't install your package in a temporary library before building the
\code{README.Rmd}. This has the pleasant side effect that, other than the latter function, it also works for \code{.Rmd} files which aren't part of an \R package.
}
\details{
Note that for public package repositories, it's recommended to use \code{\link[devtools:build_rmd]{devtools::build_readme()}} since it ensures the \code{README.Rmd} can be built \emph{reproducibly},
which means all the objects and files it references must be accessible from the repository.

This function is also registered as an \href{https://rstudio.github.io/rstudioaddins/}{RStudio add-in}, allowing RStudio users to assign a \href{https://support.rstudio.com/hc/en-us/articles/206382178-Customizing-Keyboard-Shortcuts-in-the-RStudio-IDE}{custom shortcut} to it or to invoke it from the \href{https://www.rstudio.com/blog/rstudio-1-4-a-quick-tour/#command-palette--shortcuts}{command palette}.
}
\seealso{
Other R Markdown and knitr functions:
\code{\link{knitr_table_format}()},
\code{\link{strip_yaml_header}()}
}
\concept{rmd_knitr}
