% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{build_readme}
\alias{build_readme}
\title{Build \code{README.Rmd}}
\usage{
build_readme(input = "README.Rmd", output = "README.md")
}
\arguments{
\item{input}{The path to the R Markdown README file to be built. A character scalar.}

\item{output}{The path of the built Markdown README. A character scalar.}
}
\description{
This function is an preliminary replacement for \code{\link[devtools:build_readme]{devtools::build_readme()}} that \emph{works} with the \code{\link[pal:gitlab_document]{pal::gitlab_document}} R Markdown
output format.
}
