% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{build_readme}
\alias{build_readme}
\title{Build \code{README.Rmd}}
\usage{
build_readme(
  input = "README.Rmd",
  output = "README.md",
  build_index_md = NULL,
  env = parent.frame()
)
}
\arguments{
\item{input}{The path to the R Markdown README file to be built. A character scalar.}

\item{output}{The path of the built Markdown README. A character scalar.}

\item{build_index_md}{Whether to build a separate \link[pkgdown:pkgdown-package]{pkgdown}-optimized \code{pkgdown/index.md} alongside \code{output}. If \code{NULL}, it will only be
built if the parent directory of \code{output} \link[=is_pkgdown_dir]{contains a pkgdown configuration file}.}

\item{env}{Environment in which code chunks are to be evaluated, e.g. \code{\link[=parent.frame]{parent.frame()}}, \code{\link[=new.env]{new.env()}}, or \code{\link[=globalenv]{globalenv()}}.}
}
\value{
The path to \code{input} as a character scalar, invisibly.
}
\description{
A simpler, but considerably faster alternative to \code{\link[devtools:build_rmd]{devtools::build_readme()}} since it doesn't install your package in a temporary library before building the
\code{README.Rmd}. This has the pleasant side effect that, other than the latter function, it also works for \code{.Rmd} files which aren't part of an \R package.
}
\details{
Note that for public package repositories, it's recommended to use \code{\link[devtools:build_rmd]{devtools::build_readme()}} since it ensures the \code{README.Rmd} can be built \emph{reproducibly},
which means all the objects and files it references must be accessible from the repository.

This function is also registered as an \href{https://rstudio.github.io/rstudioaddins/}{RStudio add-in} allowing RStudio users to assign a custom shortcut to it.
}
\seealso{
Other R Markdown and knitr functions:
\code{\link{gitlab_document}()}
}
\concept{rmd_knitr}
