% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{capture_print}
\alias{capture_print}
\title{Capture printed console output as string}
\usage{
capture_print(x)
}
\arguments{
\item{x}{The \R object of which the output of \code{print(x)} is to be captured.}
}
\value{
A character scalar.
}
\description{
Returns what \code{\link[base:print]{print(x)}} would output on the console as an atomic character vector (aka string).
}
\details{
This is a simple convenience wrapper around \code{\link[utils:capture.output]{utils::capture.output()}}. Note that \href{https://en.wikipedia.org/wiki/ANSI_escape_code}{ANSI escape sequences}
(e.g. as output by the \code{print()} methods of tidyverse packages) are not captured (i.e. lost).
}
\examples{
mtcars \%>\% pal::capture_print() \%T>\% str() \%>\% cat()
}
