% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{cat_lines}
\alias{cat_lines}
\title{Print \code{x} as newline-separated character vector using \code{\link[base:cat]{cat()}}.}
\usage{
cat_lines(x)
}
\arguments{
\item{x}{A vector to print.}
}
\value{
None (invisible \code{NULL}).
}
\description{
This is simply a convenience wrapper around \code{\link[base:cat]{cat()}}, mainly intended for interactive use.
}
\examples{
library(magrittr)

fs::path_package(package = "pal",
                 "DESCRIPTION") \%>\%
  readr::read_lines() \%>\%
  cat_lines()
}
