% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{cat_lines}
\alias{cat_lines}
\title{Print \code{x} as newline-separated character vector using \code{cat()}.}
\usage{
cat_lines(x)
}
\arguments{
\item{x}{A vector to print.}
}
\value{
None (invisible \code{NULL}).
}
\description{
This is simply a convenience wrapper around \code{\link[base:cat]{cat()}}, mainly intended for interactive use.
}
\examples{
fs::path_package(package = "pal",
                 "DESCRIPTION") \%>\%
  readr::read_lines() \%>\%
  cat_lines()
}
\seealso{
\code{\link[xfun:raw_string]{xfun::raw_string()}}, \code{\link[xfun:file_string]{xfun::file_string()}}

String functions
\code{\link{as_chr}()},
\code{\link{as_string}()},
\code{\link{dsv_colnames}()},
\code{\link{fuse_regex}()}
}
\concept{string}
