% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{desc_value}
\alias{desc_value}
\title{Get the value from a DESCRIPTION file field, cleaned up and with fallback}
\usage{
desc_value(key, file = ".")
}
\arguments{
\item{key}{The field to query.}

\item{file}{DESCRIPTION file to use. By default the DESCRIPTION
file of the current package (i.e. the package the working directory
is part of) is used.}
}
\value{
A character scalar.
}
\description{
Get the value from a DESCRIPTION file field, cleaned up and with fallback
}
\examples{
desc_value(key = "Description",
           file = fs::path_package("pal"))
}
