% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{escape_lf}
\alias{escape_lf}
\title{Escape line feeds / newlines}
\usage{
escape_lf(x, escape_cr = FALSE)
}
\arguments{
\item{x}{A character vector.}

\item{escape_cr}{Whether or not to also escape the carriage return character \code{CR} (aka \verb{\\r}). A logical scalar.}
}
\value{
A character vector of the same length as \code{x}.
}
\description{
This function escapes the \href{https://en.wikipedia.org/wiki/Newline}{POSIX-standard newline control character \code{LF}} (aka \verb{\\n}) which is the standard on
Unix/Linux and recent versions of macOS. Set \code{escape_cr = TRUE} in order to also escape the carriage return character \code{CR} (aka \verb{\\r}) commonly used on
Microsoft Windows.
}
\examples{
# read in and print package description as-is
text <-
  fs::path_package(package = "pal",
                   "DESCRIPTION") \%>\%
  readr::read_file() \%T>\%
  cat_lines()

# escape newlines and print again
escape_lf(text) \%>\% cat_lines()
}
\seealso{
String functions
\code{\link{as_chr}()},
\code{\link{as_string}()},
\code{\link{cat_lines}()},
\code{\link{dsv_colnames}()},
\code{\link{fuse_regex}()},
\code{\link{prose_ls}()},
\code{\link{str_replace_file}()},
\code{\link{str_replace_verbose}()}
}
\concept{string}
