% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{fuse_regex}
\alias{fuse_regex}
\title{Fuse regular expressions}
\usage{
fuse_regex(..., .literal = FALSE)
}
\arguments{
\item{...}{Regular expressions. All elements will be converted to type character before fusing. \link[rlang:dyn-dots]{Dynamic dots} are supported.}

\item{.literal}{Whether or not the input should be interpreted \emph{literally} instead of as regular expressions. If \code{TRUE}, all elements of \code{...} are wrapped in
the literal escape sequences \verb{\\Q} and \verb{\\E}.}
}
\value{
A character scalar.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}} \cr
It's recommended to rely on the \href{https://rex.r-lib.org/}{\strong{rex}} package instead, more precisely \code{\link[rex:or]{rex::rex(or(...))}} and
\code{\link[rex:escape]{rex::rex(or(rex::escape(c(...))))}}.

Combines a vector or list of regular expressions to a single one (by logical OR).
}
\examples{
stringr::str_subset(string = c("yeah",
                               "nope",
                               "maybe",
                               "never"),
                    pattern = pal::fuse_regex(letters[1:3]))
}
\seealso{
The \link[rex:rex]{rex} package which provides an intuitive framework to build complex regular expressions.

Other string functions:
\code{\link{as_chr}()},
\code{\link{as_comment_str}()},
\code{\link{as_env_var_name}()},
\code{\link{as_line_feed_chr}()},
\code{\link{as_str}()},
\code{\link{capitalize_first}()},
\code{\link{dsv_colnames}()},
\code{\link{enum_str}()},
\code{\link{escape_lf}()},
\code{\link{phrase_nr}()},
\code{\link{sentenceify}()},
\code{\link{wrap_chr}()}
}
\concept{string}
