% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{fuse_regex}
\alias{fuse_regex}
\title{Fuse regular expressions}
\usage{
fuse_regex(...)
}
\arguments{
\item{...}{The regular expressions. All elements will be converted to type character before fusing.}
}
\value{
A character scalar.
}
\description{
Combine a vector or list of regular expressions to a single one (by logical OR).
}
\examples{
\donttest{
# perform some (nonsense) Jane Austen text extraction
regex <- c("My dear Jane",
           "make haste, ",
           "(?i)\\\\bevil")

stringr::str_subset(string = janeaustenr::prideprejudice,
                    pattern = fuse_regex(regex))}
}
\seealso{
String functions
\code{\link{as_chr}()},
\code{\link{as_string}()},
\code{\link{cat_lines}()},
\code{\link{dsv_colnames}()},
\code{\link{escape_lf}()},
\code{\link{prose_ls}()},
\code{\link{str_replace_file}()},
\code{\link{str_replace_verbose}()}
}
\concept{string}
