% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{git_remote_tree_url}
\alias{git_remote_tree_url}
\title{Get Git remote tree URL}
\usage{
git_remote_tree_url(repo = ".")
}
\arguments{
\item{repo}{The path to the git repository. If the directory is not a
repository, parent directories are considered (see \link[gert]{git_find}). To disable
this search, provide the filepath protected with \code{\link[=I]{I()}}. When using this
parameter, always explicitly call by name (i.e. \verb{repo = }) because future
versions of gert may have additional parameters.}
}
\value{
If \code{repo} is a Git repository, a character scalar. Otherwise, a character vector of length zero.
}
\description{
Determines the base Git tree URL to the current branch's upstream remote of \code{repo}.
}
\details{
This function is useful to assemble URLs to files and folders in your repo's Git forge (GitHub, GitLab, etc.).
}
\examples{
pal::git_remote_tree_url() |>
  paste0("Rmd/pal.Rmd") |>
  browseURL()
}
\seealso{
Other Git functions:
\code{\link{git_file_mod_time}()}
}
\concept{git}
