% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{is_http_success}
\alias{is_http_success}
\title{Test if an HTTP request is successful}
\usage{
is_http_success(url)
}
\arguments{
\item{url}{The HTTP protocol address. The scheme is optional, so both \code{"google.com"} and \code{"https://google.com"} will work. A character scalar.}
}
\value{
A logical scalar.
}
\description{
This is a convenience wrapper around \code{\link[httr:http_error]{!httr::http_error()}} that returns
\itemize{
\item \code{TRUE} if the specified \code{url} could be resolved \emph{and} a \href{https://en.wikipedia.org/wiki/Hypertext_Transfer_Protocol#Request_methods}{\code{HEAD} request} could
be \href{https://en.wikipedia.org/wiki/List_of_HTTP_status_codes}{successfully completed}, or
\item \code{FALSE} in any other case.
}
}
\details{
This function is similar to \code{\link[RCurl:url.exists]{RCurl::url.exists()}}, i.e. it only retrieves the header, no body, but is based on \link[httr:httr-package]{httr} which in turn is
based on \href{https://jeroen.cran.dev/curl/}{curl}.
}
\examples{
is_http_success("goo.gl")
is_http_success("https://google.com/")
is_http_success("https://google.not/")
}
\seealso{
\code{\link[RCurl:url.exists]{RCurl::url.exists()}}
}
