% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{is_pkg_cran}
\alias{is_pkg_cran}
\title{Test if a package is available on CRAN}
\usage{
is_pkg_cran(pkg, min_version = NULL, retries = 1L)
}
\arguments{
\item{pkg}{Package name (case-sensitive). A character scalar.}

\item{min_version}{Minimum required version number of \code{pkg}. Must be either \code{NULL} to ignore version numbers, or a single
\code{\link[base:numeric_version]{package_version}} or something coercible to.}

\item{retries}{Maximum number of retries of the \code{HEAD} request in case of an HTTP error. An integer scalar \verb{>= 0}. The retries are performed using
exponential backoff and jitter, see \code{\link[httr:RETRY]{httr::RETRY()}} for details.}
}
\value{
A logical scalar.
}
\description{
Test if a package is available on CRAN
}
\examples{
pal::is_pkg_cran("foobar")
pal::is_pkg_cran("dplyr")
pal::is_pkg_cran("dplyr", min_version = 9999.9)
}
\seealso{
Other R package functions:
\code{\link{assert_pkg}()},
\code{\link{augment_pkg_config}()},
\code{\link{has_pkg_config_val}()},
\code{\link{is_pkg_dir}()},
\code{\link{is_pkg_installed}()},
\code{\link{is_pkgdown_dir}()},
\code{\link{ls_pkg}()},
\code{\link{pkg_config_val}()},
\code{\link{use_pkg}()}
}
\concept{rpkgs}
