% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{is_pkg_dir}
\alias{is_pkg_dir}
\title{Test if a directory is an \R package}
\usage{
is_pkg_dir(path = ".")
}
\arguments{
\item{path}{Path of the directory to check. A character scalar. Defaults to the current working directory.}
}
\value{
\code{TRUE} if \code{path} is the root directory of an \R package, \code{FALSE} otherwise.
}
\description{
Convenience wrapper around the \code{\link[rprojroot:criteria]{rprojroot::is_r_package}} root criterion. Note that it will only return \code{TRUE} for the root of a
package directory, not its subdirectories.
}
\examples{
pal::is_pkg_dir()
pal::is_pkg_dir(fs::path_package("utils"))
}
\seealso{
Other R package functions:
\code{\link{assert_pkg}()},
\code{\link{is_pkg_cran}()},
\code{\link{is_pkg_installed}()},
\code{\link{is_pkgdown_dir}()},
\code{\link{ls_pkg}()},
\code{\link{use_pkg}()}
}
\concept{rpkgs}
