% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{is_pkg_dir}
\alias{is_pkg_dir}
\title{Test if a directory is an R package}
\usage{
is_pkg_dir(path = ".")
}
\arguments{
\item{path}{The path of the directory to check. A character scalar. Defaults to the current working directory.}
}
\value{
\code{TRUE} if \code{path} is the root directory of an R package, \code{FALSE} otherwise.
}
\description{
This is a simple convenience wrapper around the \code{\link[rprojroot:criteria]{rprojroot::is_r_package}} root criterion. Note that this function will only
return \code{TRUE} for the root of a package directory, not its subdirectories.
}
\examples{
is_pkg_dir()
is_pkg_dir(fs::path_package("pal"))
}
\seealso{
R package functions
\code{\link{is_installed}()},
\code{\link{ls_pkg}()}
}
\concept{rpkgs}
