% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{is_pkg_installed}
\alias{is_pkg_installed}
\title{Test if packages are installed}
\usage{
is_pkg_installed(pkg, min_version = NULL)
}
\arguments{
\item{pkg}{Package names. A character vector.}

\item{min_version}{Minimum required version number of each \code{pkg}. Must be either \code{NULL} to ignore version numbers, or a vector of
\code{\link[base:numeric_version]{package_version}}s or something coercible to.}
}
\value{
A named logical vector of the same length as \code{pkg}.
}
\description{
Returns \code{TRUE} or \code{FALSE} for each \code{pkg}, depending on whether the \code{pkg} is installed on the current system or not, optionally ensuring a \code{min_version}.
}
\details{
In contrast to \code{\link[base:library]{base::require()}}, this function checks if the packages are installed without attaching their namespaces if so.

In contrast to \code{\link[rlang:is_installed]{rlang::is_installed()}} or \code{\link[xfun:pkg_attach]{xfun::pkg_available()}}, this function doesn't load the packages if they're installed.

In contrast to \code{\link[xfun:pkg_attach]{xfun::pkg_available()}}, it is fully vectorized, i.e. returns a (named) logical vector of the same length as \code{pkg}.

It is \href{https://stackoverflow.com/questions/9341635/check-for-installed-packages-before-running-install-packages/38082613#38082613}{considerably faster} than the commonly used
\code{pkg \%in\% rownames(installed.packages())} check.
}
\examples{
pal::is_pkg_installed("tidyverse")

# it is vectorized...
pal::is_pkg_installed(pkg = c("dplyr", "tibble", "magrittr"),
                      min_version = c("1.0", "2.0", "99.9.9000"))

# ...and scalar arguments will be recycled
pal::is_pkg_installed(pkg = "dplyr",
                      min_version = c("0.5", "1.0", "99.9.9000"))

pal::is_pkg_installed(pkg = c("dplyr", "tibble", "magrittr"),
                      min_version = "1.0")
}
\seealso{
Other R package functions:
\code{\link{is_pkg_cran}()},
\code{\link{is_pkg_dir}()},
\code{\link{is_pkgdown_dir}()},
\code{\link{ls_pkg}()},
\code{\link{use_pkg}()}
}
\concept{rpkgs}
