% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{ls_pkg}
\alias{ls_pkg}
\title{List a subset of all installed packages}
\usage{
ls_pkg(pkg, ignore_case = TRUE)
}
\arguments{
\item{pkg}{A character vector of package names.}

\item{ignore_case}{If \code{FALSE}, \code{pkg} is case-sensitive.}
}
\value{
A \link[tibble:tbl_df]{tibble}.
}
\description{
List a subset of all installed packages
}
\examples{
ls_pkg(pkg = c("pal", "tibble", "dplyr"))
}
\seealso{
R package functions
\code{\link{is_installed}()},
\code{\link{is_pkg_dir}()}
}
\concept{rpkgs}
