% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{md_xml_subnode_ix}
\alias{md_xml_subnode_ix}
\title{Determine CommonMark XML subnode indices}
\usage{
md_xml_subnode_ix(xml)
}
\arguments{
\item{xml}{CommonMark parse tree. An \code{\link[xml2:oldclass]{xml_document}}, \code{\link[xml2:oldclass]{xml_nodeset}} or
\code{\link[xml2:oldclass]{xml_node}}.}
}
\value{
A list of integer vectors of the same length as the number of XML nodes at the highest level of \code{xml}.
}
\description{
Determines the XML children node indices for every XML node at the highest level of \code{xml} by interpreting \href{https://pandoc.org/MANUAL.html#headings}{Markdown heading levels} (1–6).
}
\details{
\code{\link[commonmark:commonmark]{commonmark::markdown_xml()}} (and so \code{\link[=md_to_xml]{md_to_xml()}} which builds upon it) \strong{do} parse (R) Markdown file content according to the
\strong{\href{https://commonmark.org/}{CommonMark} specification}, but \strong{do not} return any information about the document's \strong{heading hierarchy}.
\code{md_xml_subnode_ix()} fills this gap by giving the hierarchy structure in the form of the XML subnode indices for every node at the highest level of \code{xml}.
}
\examples{
"# A title

Some prose.

## A subtitle

More prose.

## Another subtitle

Out of prose here.

### A sub-subtitle

I'm dug in.

# Another title

A last word." |>
  pal::md_to_xml() |>
  pal::md_xml_subnode_ix()
}
\seealso{
Other CommonMark parsing functions:
\code{\link{md_to_xml}()},
\code{\link{xml_to_md}()}
}
\concept{commonmark}
