% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\docType{package}
\name{pal-package}
\alias{pal}
\alias{pal-package}
\title{pal: A color palette search engine.}
\description{
\if{html}{\figure{logo.png}{options: align='right' width = 100px}}

Search the color palette that best suit
    your interests can become a tired activity.
    This package facilitates find your ideal color
    palette by making the package collections: cpt-city,
    hex-color and paletteer more accessible to realize
    queries. There are more than 80 000 > palettes
    waiting to be used!.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/csaybar/pal/}
  \item Report bugs at \url{https://github.com/csaybar/pal/issues}
}

}
\keyword{internal}
