% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-package.R
\docType{package}
\name{pal-package}
\alias{pal-package}
\title{pal: LLM-driven coding assistants}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

LLM-driven coding assistants.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/pal}
  \item \url{https://simonpcouch.github.io/pal/}
  \item Report bugs at \url{https://github.com/simonpcouch/pal/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

}
\keyword{internal}
