% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-package.R
\docType{package}
\name{pal-package}
\alias{pal}
\alias{pal-package}
\title{pal: LLM assistants for R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a library of ergonomic LLM assistants designed to help you complete repetitive, hard-to-automate tasks quickly. After selecting some code, press the keyboard shortcut you've chosen to trigger the pal addin, select a pal, and watch your code be rewritten. While a number of pals for R package development ship with the package, users can create a suite of custom pals with only a directory of markdown files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/pal}
  \item \url{https://simonpcouch.github.io/pal/}
  \item Report bugs at \url{https://github.com/simonpcouch/pal/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
