% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{pal}
\alias{pal}
\title{Create a pal}
\usage{
pal(
  role = NULL,
  keybinding = NULL,
  fn = getOption(".pal_fn", default = "new_chat_claude"),
  ...,
  .ns = "elmer"
)
}
\arguments{
\item{role}{The identifier for a pal prompt. Currently one
of \link[=pal_cli]{"cli"} or \link[=pal_testthat]{"testthat"}.}

\item{keybinding}{A key binding for the pal. \strong{Currently unused.}
Keybdings have to be registered in the usual way (via Tools >
Modify Keyboard Shortcuts), for now.}

\item{fn}{A \verb{new_*()} function, likely from the elmer package. Defaults
to \code{\link[elmer:new_chat_claude]{elmer::new_chat_claude()}}. To set a persistent alternative default,
set the \code{.pal_fn} option; see examples below.}

\item{...}{Additional arguments to \code{fn}. The \code{system_prompt} argument will
be ignored if supplied. To set persistent defaults,
set the \code{.pal_args} option; see examples below.}

\item{.ns}{The package that the \verb{new_*()} function is exported from.}
}
\description{
Pals are persistent, ergonomic LLM assistants designed to help you complete
repetitive, hard-to-automate tasks quickly. When created, they automatically
generate RStudio add-ins registered to keyboard shortcuts. After selecting
some code, press the keyboard shortcut you've chosen and watch your code
be rewritten.
}
\details{
Upon successfully creating a pal, this function will assign the
result to the search path as \code{.last_pal}.

If you have an Anthropic API key (or another API key and the \verb{pal_*()}
options) set and this package installed, you are ready to using the add-in
in any R session with no setup or library loading required; the addin knows
to look for your API credentials and will call needed functions by itself.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# to create a chat with claude:
pal()

# or with OpenAI's 4o-mini:
pal(
  "new_chat_openai",
  model = "gpt-4o-mini"
)

# to set OpenAI's 4o-mini as the default, for example, set the
# following options (possibly in your .Rprofile, if you'd like
# them to persist across sessions):
options(
  .pal_fn = "new_chat_openai",
  .pal_args = list(model = "gpt-4o-mini")
)
\dontshow{\}) # examplesIf}
}
