% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-add-remove.R
\name{pal_add_remove}
\alias{pal_add_remove}
\alias{.pal_add}
\alias{.pal_remove}
\title{Creating custom pals}
\usage{
.pal_add(role, prompt = NULL, interface = c("replace", "prefix", "suffix"))

.pal_remove(role)
}
\arguments{
\item{role}{A single string giving a descriptor of the pal's functionality.}

\item{prompt}{A file path to a markdown file giving the system prompt or
the output of \code{\link[elmer:interpolate]{elmer::interpolate()}}.}

\item{interface}{One of \code{"replace"}, \code{"prefix"}, or \code{"suffix"}, describing
how the pal will interact with the selection. For example, the
\link[=pal_cli]{cli pal} \code{"replace"}s the selection, while the
\link[=pal_roxygen]{roxygen pal} \code{"prefixes"} the selected code with documentation.}

\item{dir}{A directory of markdown files. See "Adding multiple, persistent
pals" section below.}
}
\value{
\code{NULL}, invisibly. Called for its side effect: a pal with role \code{role}
is registered with the pal package.
}
\description{
Users can create custom pals using the \code{.pal_add()} function; after passing
the function a role and prompt, the pal will be available on the command
palette.
}
\section{Adding multiple, persistent pals}{

}

