% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-add-remove.R
\name{pal_add_remove}
\alias{pal_add_remove}
\alias{.pal_add}
\alias{.pal_remove}
\title{Registering pals}
\usage{
.pal_add(role, prompt = NULL, interface = c("replace", "prefix", "suffix"))

.pal_remove(role)
}
\arguments{
\item{role}{A single string giving a descriptor of the pal's functionality.
Cand only contain letters and numbers.}

\item{prompt}{A single string giving the system prompt. In most cases, this
is a rather long string, containing several newlines.}

\item{interface}{One of \code{"replace"}, \code{"prefix"}, or \code{"suffix"}, describing
how the pal will interact with the selection. For example, the
\link[=pal_cli]{cli pal} \code{"replace"}s the selection, while the
\link[=pal_roxygen]{roxygen pal} \code{"prefixes"} the selected code with documentation.}
}
\value{
\code{NULL}, invisibly. Called for its side effect: a pal with role \code{role}
is registered (or unregistered) with the pal package.
}
\description{
Users can create custom pals using the \code{.pal_add()} function; after passing
the function a role and prompt, the pal will be available in the
\link[=.init_addin]{pal addin}.

\strong{Most users should not need to interact with these functions.}
\code{\link[=prompt_new]{prompt_new()}} and friends can be used to create prompts for new pals, and
those pals can be registered with pal using \code{\link[=directory_load]{directory_load()}} and friends.
The pals created by those functions will be persistent across sessions.
}
