% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pal_paletteer}
\alias{pal_paletteer}
\title{pal_paletteer: Comprehensive Collection of Color Palettes}
\format{In This data_frame each row represent a single color palette,
the columns contains the following: \cr

\itemize{
\item tags: a character vector; label attached used for identification purpose.
\item color: a list of character vectors; Contains color information in hexadecimal notation.
\item n:  Numbers of colors used in the palette.
\item matrix: a list of matrices; Contains color information at CIELAB color space.
\item type: Type of the pallete. The type avaible are "discrete" or "continuous".
\item alpha: logical; TRUE/FALSE whether the palette present at least color with transparent.
\item names: character; "key" used for palette identification.
}}
\usage{
pal_paletteer
}
\description{
(Modified of this \href{https://emilhvitfeldt.github.io/paletteer/}{website})
Paletteer is a great R package that collect all color palettes
across the R ecosystem  under the same package with a
streamlined API.
}
\note{
pal_paletteer comprises 1737 palettes. The author and mantainer
is \href{https://github.com/EmilHvitfeldt/}{Emil Hvitfeldt}.
For get more information about the authors' copyright,
check the following \href{https://github.com/EmilHvitfeldt/paletteer/blob/master/LICENSE.note}{license} file.
}
\keyword{datasets}
