% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_search.R
\name{pal_search}
\alias{pal_search}
\title{Create a pal search}
\usage{
pal_search(catalogs = c("cpt-city", "color-hex", "paletteer"),
  type = c("continuous", "discrete", "diverging", "qualitative",
  "sequential"), return = 25)
}
\arguments{
\item{catalogs}{character vector; Collections of color palette.
The color palette collections used in pal package come from: \cr
\itemize{
\item cpt-city: http://soliton.vm.bytemark.co.uk/pub/cpt-city/index.html \cr
\item color-hex: https://www.color-hex.com/color-palettes/
\item paletteer: https://github.com/EmilHvitfeldt/paletteer
}}

\item{type}{character vector; Type of palette. The type: \cr
\itemize{
\item 'diverging' \cr
\item 'qualitative' \cr
\item 'sequential' \cr
only are supported by the 'paletteer'catalogs.
}}

\item{return}{numeric; Specify the number of palettes to be returned.}
}
\description{
pal_search() initializes a pal object. It is
used to set the global parameters of 'pal' search.
}
