% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{path_mod_time}
\alias{path_mod_time}
\title{Get path modification time}
\usage{
path_mod_time(path, follow = FALSE)
}
\arguments{
\item{path}{Path to a file or directory. A character scalar.}

\item{follow}{If \code{TRUE}, symbolic links will be followed (recursively) and
the results will be that of the final file rather than the link.}
}
\value{
A \link[base:DateTimeClasses]{datetime}.
}
\description{
Determine a file's or directory's modification time. Other than \code{\link[=file.mtime]{file.mtime()}} datetime of modification is always returned in
\href{https://en.wikipedia.org/wiki/Coordinated_Universal_Time}{UTC}.
}
\examples{
fs::path_package(package = "pal",
                 "DESCRIPTION") |>
  pal::path_mod_time()
}
