% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{path_script}
\alias{path_script}
\title{Determine file path of executing script}
\usage{
path_script()
}
\value{
The file path to the executing script.
}
\description{
Tries to determine the path to the R/Rmd script that this function is called from.
}
\seealso{
Other system interaction functions:
\code{\link{assert_cli}()},
\code{\link{test_cli}()}
}
\concept{sys}
