% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{pipe_table}
\alias{pipe_table}
\title{Convert dataframe/tibble to Markdown pipe table}
\usage{
pipe_table(
  x,
  strong_colnames = TRUE,
  strong_rownames = TRUE,
  justify = "left",
  ...
)
}
\arguments{
\item{x}{The dataframe/tibble/matrix to be converted to a pipe table.}

\item{strong_colnames}{Highlight column names by formatting them \verb{<strong>}. Enabled by default.}

\item{strong_rownames}{Highlight row names by formatting them \verb{<strong>}. Enabled by default.}

\item{justify}{defines alignment in cells passed to \code{format}. Can be \code{left}, \code{right} or \code{centre}, which latter can be also spelled as \code{center}. Defaults to \code{centre}. Can be abbreviated to a string consisting of the letters \code{l}, \code{c} and \code{r} (e.g. 'lcr' instead of c('left', 'centre', 'right').}

\item{...}{Additional arguments passed to \code{\link[pander:pandoc.table.return]{pander::pandoc.table.return()}}.}
}
\value{
A character scalar.
}
\description{
This is a simple wrapper around \code{\link[pander:pandoc.table.return]{pander::pandoc.table.return()}} with sensible defaults to create a
\href{https://pandoc.org/MANUAL.html#extension-pipe_tables}{Markdown pipe table}.
}
\section{Create tables dynamically in roxygen2 documentation}{
This function can be useful to create tables inside \link[roxygen2:roxygen2]{roxygen2} documentation programmatically from data using
\href{https://roxygen2.r-lib.org/articles/rd-formatting.html#dynamic-r-code-1}{dynamic R code}.

The inline code

\code{`r mtcars \%>\% head() \%>\% pal::pipe_table()`}

should produce the following table in \href{https://www.tidyverse.org/blog/2020/03/roxygen2-7-1-0/}{roxygen2 7.1.0} and above:\tabular{llllllllllll}{
     \tab \strong{mpg} \tab \strong{cyl} \tab \strong{disp} \tab \strong{hp} \tab \strong{drat} \tab \strong{wt} \tab \strong{qsec} \tab \strong{vs} \tab \strong{am} \tab \strong{gear} \tab \strong{carb} \cr
   \strong{Mazda RX4} \tab 21 \tab 6 \tab 160 \tab 110 \tab 3.9 \tab 2.62 \tab 16.46 \tab 0 \tab 1 \tab 4 \tab 4 \cr
   \strong{Mazda RX4 Wag} \tab 21 \tab 6 \tab 160 \tab 110 \tab 3.9 \tab 2.875 \tab 17.02 \tab 0 \tab 1 \tab 4 \tab 4 \cr
   \strong{Datsun 710} \tab 22.8 \tab 4 \tab 108 \tab 93 \tab 3.85 \tab 2.32 \tab 18.61 \tab 1 \tab 1 \tab 4 \tab 1 \cr
   \strong{Hornet 4 Drive} \tab 21.4 \tab 6 \tab 258 \tab 110 \tab 3.08 \tab 3.215 \tab 19.44 \tab 1 \tab 0 \tab 3 \tab 1 \cr
   \strong{Hornet Sportabout} \tab 18.7 \tab 8 \tab 360 \tab 175 \tab 3.15 \tab 3.44 \tab 17.02 \tab 0 \tab 0 \tab 3 \tab 2 \cr
   \strong{Valiant} \tab 18.1 \tab 6 \tab 225 \tab 105 \tab 2.76 \tab 3.46 \tab 20.22 \tab 1 \tab 0 \tab 3 \tab 1 \cr
}
}

\examples{
library(magrittr)
mtcars \%>\% head() \%>\% pal::pipe_table() \%>\% cat()
}
