% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{pkg_config_val}
\alias{pkg_config_val}
\title{Get package configuration value}
\usage{
pkg_config_val(key, pkg)
}
\arguments{
\item{key}{Configuration key name. A character scalar.}

\item{pkg}{Package name. A character scalar.}
}
\value{
\code{pkgsnip::return_label("r_obj")}
}
\description{
Retrieves a package configuration value in a canonical way. The following configuration sources are consulted in descending order and the first hit wins:
\enumerate{
\item The \R option \verb{<pkg>.<key>}.
\item The environment variable \verb{<PKG>_<KEY>}.
\item The configuration's default value as specified in the package's configuration metadata (\verb{<pkg>::pkg_config$default_value}). If no default value is
specified (\code{NULL}), an error is thrown.
}
}
\details{
This function is intended to be used by package authors who want to expose their package configuration options in a canonical way (as outlined above). For
\code{pkg_config_val()} to properly work, the configuration metadata must be available in the package's namespace as object \code{pkg_config}, which must be a
\link[=data.frame]{dataframe} or \link[tibble:tbl_df-class]{tibble} with at minimum the columns \code{key} (of type character holding the configuration key names) and
\code{default_value} (of type list holding the default configuration values, if any).
}
