% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{pkg_config_val}
\alias{pkg_config_val}
\title{Get package configuration value}
\usage{
pkg_config_val(key, pkg, default = NULL)
}
\arguments{
\item{key}{Configuration key name. A character scalar.}

\item{pkg}{Package name. A character scalar.}

\item{default}{Default value to fall back to if neither the \R option \verb{<pkg>.<key>} nor the environment variable \verb{<PKG>_<KEY>} is set. If \code{NULL}, the
default value for \code{key} in \verb{<pkg>::pkg_config} will be used (if defined).}
}
\value{
An \R object.
}
\description{
Retrieves a package configuration value in a canonical way. The following configuration sources are consulted in descending order and the first hit is
returned:
\enumerate{
\item The \R \link[=options]{option} \verb{<pkg>.<key>}.
\item The \href{https://en.wikipedia.org/wiki/Environment_variable}{environment variable} \verb{<PKG>_<KEY>}.
\item The ad-hoc default value specified via this function's \code{default} argument (\code{NULL} means unspecified).
\item The configuration's global default value as specified in the package's configuration metadata (\verb{<pkg>::pkg_config$default_value}). If no default value is
specified (\code{NULL}), an error is thrown.
}
}
\details{
This function is intended to be used by package authors who want to expose their package configuration options in a canonical way (as outlined above). For
\code{pkg_config_val()} to properly work, the configuration metadata must be available in the package's namespace as object \code{pkg_config}, which must be a
\link[=data.frame]{dataframe} or \link[tibble:tbl_df-class]{tibble} with at minimum the columns \code{key} (of type character holding the configuration key names) and
\code{default_value} (of type list holding the default configuration values, if any).
}
\examples{
try(
  pal::pkg_config_val(key = "gen_pkgdown_ref",
                      pkg = "pkgpurl")
)
}
\seealso{
\code{\link[xfun:env_option]{xfun::env_option()}} for an alternative approach to R option and environment variable coherence.

Other package configuration functions:
\code{\link{augment_pkg_config}()},
\code{\link{has_pkg_config_val}()},
\code{\link{pkg_config_val_default}()},
\code{\link{print_pkg_config}()}
}
\concept{pkg_config}
