% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{print_pkg_config}
\alias{print_pkg_config}
\title{Print package configuration metadata}
\usage{
print_pkg_config(pkg)
}
\arguments{
\item{pkg}{Package name. A character scalar.}
}
\value{
A character vector.
}
\description{
Prints a package's configuration metadata (\verb{<pkg>::pkg_config}) as a prettily formatted Markdown table.
}
\examples{
try(
  pal::print_pkg_config(pkg = "pkgpurl")
)
}
\seealso{
Other package configuration functions:
\code{\link{augment_pkg_config}()},
\code{\link{has_pkg_config_val}()},
\code{\link{pkg_config_val}()}
}
\concept{pkg_config}
