% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{prose_ls}
\alias{prose_ls}
\title{List items concatenated in prose style (..., ... and ...)}
\usage{
prose_ls(x, wrap = "", separator = ", ", last_separator = " and ")
}
\arguments{
\item{x}{A vector or a list.}

\item{wrap}{The string (usually a single character) in which \code{x} is to be wrapped.}

\item{separator}{The separator to delimit the elements of \code{x}.}

\item{last_separator}{The separator to delimit the second-last and last element of \code{x}.}
}
\value{
A character scalar.
}
\description{
Takes a vector or list and concatenates its elements to a single string separated in prose-style.
}
\examples{
pal::prose_ls(1:5)
}
\seealso{
Other string functions:
\code{\link{as_chr}()},
\code{\link{as_comment_string}()},
\code{\link{as_string}()},
\code{\link{cat_lines}()},
\code{\link{dsv_colnames}()},
\code{\link{escape_lf}()},
\code{\link{fuse_regex}()},
\code{\link{prettify_nr}()},
\code{\link{sentenceify}()},
\code{\link{wrap_chr}()}
}
\concept{string}
