% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{reduce_df_list}
\alias{reduce_df_list}
\title{Reduce a nested list of data frames / tibbles to a single tibble}
\usage{
reduce_df_list(x, strict = TRUE)
}
\arguments{
\item{x}{A list containing data frames / tibbles at its leafs.}

\item{strict}{Ensure \code{x} contains data frames / tibbles only and throw an error otherwise. If \code{FALSE}, leafs containing other objects are ignored (skipped).}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Recursively reduces a nested list containing data frames / tibbles at its leafs to a single tibble.
}
