% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{roxy_blocks}
\alias{roxy_blocks}
\title{Get roxygen2 blocks}
\usage{
roxy_blocks(pkg = NULL, ..., text = NULL, quiet = TRUE)
}
\arguments{
\item{pkg, text}{Either a package name (\code{pkg}) or a character vector of \R source code lines (\code{text}) to extract the object's roxygen2 tag value from.}

\item{...}{Further arguments passed on to \code{\link[utils:download.packages]{utils::download.packages()}}, excluding
\code{"pkgs"}, \code{"destdir"}, \code{"available"}, \code{"type"} and \code{"quiet"}. Only relevant if \code{pkg} is provided.}

\item{quiet}{Whether or not to suppress printing status output from internal processing.}
}
\value{
A list of \code{\link[roxygen2:roxy_block]{roxy_block}} objects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Parses the roxygen2 package documentation of a specific R package or from a single \code{.R} source code file.
}
\examples{
# Either provide an R source file as a character vector `text`...
text <- readr::read_lines(paste0("https://raw.githubusercontent.com/r-lib/rlang/",
                                 "db52a58d505b65f58ba922d4752b5b0061f2a98c/R/fn.R"))

pal::roxy_blocks(text = text) |> head(n = 3L)

# ...or provide a package name as `pkg`
try(
  pal::roxy_blocks(pkg = "tinkr",
                   repos = "https://cloud.r-project.org") |>
    head(n = 3L)
)
}
\seealso{
Other package documentation functions:
\code{\link{roxy_obj}()},
\code{\link{roxy_tag_value}()},
\code{\link{roxy_to_md_links}()}
}
\concept{roxy}
