% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{roxy_tag_value}
\alias{roxy_tag_value}
\title{Get an object's roxygen2 tag value}
\usage{
roxy_tag_value(text, obj_name, tag_name = "param", param_name = NULL)
}
\arguments{
\item{text}{The \R source code to extract the object's roxygen2 tag value from. A character vector.}

\item{obj_name}{The object name to which the roxygen2 tag belongs to, usually a function name. A character scalar.}

\item{tag_name}{The name of the \href{https://roxygen2.r-lib.org/articles/rd.html}{roxygen2 tag} (without the \code{@}) to extract the value from. A character scalar.}

\item{param_name}{The parameter name to extract the value from. Only relevant if \code{tag_name = "param"}. A character scalar.}
}
\value{
A character scalar if \code{tag_name = "param"} and \code{param_name != NULL}, otherwise a list.
}
\description{
Parses \code{text} for \link[roxygen2:roxy_block]{roxygen2 blocks} and extracts the value(s) belonging to the \code{tag_name}s documenting \code{obj_name}.
}
\examples{
text <- readr::read_lines(paste0("https://raw.githubusercontent.com/r-lib/rlang/",
                                 "db52a58d505b65f58ba922d4752b5b0061f2a98c/R/fn.R"))

pal::roxy_tag_value(text = text,
                    obj_name = "as_function",
                    param_name = "x")
}
