% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{safe_seq_len}
\alias{safe_seq_len}
\title{Generate an integer sequence of specific length (safe)}
\usage{
safe_seq_len(n)
}
\arguments{
\item{n}{Desired length of the integer sequence.}
}
\value{
An integer sequence starting from \code{1L}.
}
\description{
Modified version of \code{\link[base:seq]{base::seq_len()}} that returns a zero-length integer in case of a zero-length input instead of throwing an error.
}
\examples{
pal::safe_seq_len(5)

# this function simply returns a zero-length integer for zero-length inputs...
pal::safe_seq_len(NULL)
pal::safe_seq_len(integer())

# ...while `seq_len()` throws an error
try(seq_len(NULL))
try(seq_len(integer()))
}
\seealso{
Other statistical computing / numeric functions:
\code{\link{round_to}()},
\code{\link{safe_max}()},
\code{\link{safe_min}()},
\code{\link{stat_mode}()}
}
\concept{stat}
