% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{strip_md}
\alias{strip_md}
\title{Strip Markdown formatting from character vector}
\usage{
strip_md(x, strip_footnotes = TRUE)
}
\arguments{
\item{x}{A character vector to strip Markdown formatting from.}

\item{strip_footnotes}{Whether to remove Markdown footnotes, too. If \code{FALSE}, footnotes are canonicalized (to have sequential integer identifiers).}
}
\value{
A character vector of the same length as \code{x}.
}
\description{
Removes all Markdown formatting from a character vector.
}
\details{
This function relies on \code{\link[commonmark:commonmark]{commonmark::markdown_text()}} which \href{https://github.com/jeroen/commonmark#readme}{supports the CommonMark specification plus the Github extensions}. Unfortunately, \href{https://pandoc.org/MANUAL.html#footnotes}{Markdown footnotes} can't be stripped
using \code{commonmark::markdown_text()}. Therefore a separate option \code{strip_footnotes} is offered which relies on regular expressions to remove inline footnotes
and footnote definitions and references.
}
\examples{
pal::strip_md(
  "A **MD** formatted [string](https://en.wikipedia.org/wiki/String_(computer_science))"
)

# link references are only removed *iff* the reference is included in `x`:
pal::strip_md("[A reference link][refid]\n\n[refid]: https://example.com")
pal::strip_md("[A reference link][refid]\n\n_No ref here..._")
}
\seealso{
Other (Pandoc) Markdown functions:
\code{\link{as_md_list}()},
\code{\link{as_md_val_list}()},
\code{\link{as_md_vals}()},
\code{\link{md_verb}()},
\code{\link{pipe_table}()},
\code{\link{strip_md_footnotes}()}
}
\concept{md}
