% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{toml_read}
\alias{toml_read}
\title{Read in and parse TOML file as strict list}
\usage{
toml_read(path, verbose = FALSE)
}
\arguments{
\item{path}{Path to a TOML file. A character scalar.}

\item{verbose}{A logical switch to turn on (very) verbose operation
    which can be useful in debugging a parsing issue in a file.}
}
\value{
A \link[xfun:strict_list]{strict list}.
}
\description{
Reads in a file in \href{https://toml.io/}{Tom's Obvious Minimal Language (TOML)} format and returns its content as a (nested) \link[xfun:strict_list]{strict list}.
}
\details{
The file is parsed using \code{\link[RcppTOML:parseTOML]{RcppTOML::parseTOML(escape = FALSE)}}.
}
\seealso{
Other TOML functions:
\code{\link{toml_validate}()}
}
\concept{toml}
