% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-GEN.R
\name{wrap_chr}
\alias{wrap_chr}
\title{Wrap character vector in string}
\usage{
wrap_chr(x, wrap = "\\"")
}
\arguments{
\item{x}{A character vector or something coercible to. Will be fed to \code{\link[=as_chr]{as_chr()}} before wrapping.}

\item{wrap}{The string \code{x} is to be wrapped in.}
}
\value{
A character vector.
}
\description{
Wrap character vector in string
}
\examples{
mtcars \%>\%
  magrittr::set_colnames(wrap_chr(x = colnames(.),
                                  wrap = "`"))
  \%>\% pipe_table
}
