% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.gen.R
\name{xml_to_md}
\alias{xml_to_md}
\title{Convert from CommonMark XML to (R) Markdown}
\usage{
xml_to_md(xml)
}
\arguments{
\item{xml}{CommonMark parse tree. An \code{\link[xml2:oldclass]{xml_document}}, \code{\link[xml2:oldclass]{xml_nodeset}} or
\code{\link[xml2:oldclass]{xml_node}}.}
}
\value{
A character scalar.
}
\description{
Convert from CommonMark XML to (R) Markdown
}
\examples{
"# A title

Some prose.

## A subtitle

More prose.

## Another subtitle

Out of prose here.

### A sub-subtitle

I'm dug in.

# Another title

A last word." |>
  pal::md_to_xml() |>
  xml2::xml_contents() |>
  pal::xml_to_md() |>
  cat()
}
\seealso{
Other CommonMark parsing functions:
\code{\link{md_to_xml}()},
\code{\link{md_xml_subnode_ix}()}
}
\concept{commonmark}
